refactor final version need to be tested
