+++ orgs.go
@@ -1,4 +1,47 @@
-org.go
+package shuffle
+
+import (
+	"bytes"
+	"context"
+	"errors"
+	"fmt"
+	"gopkg.in/yaml.v3"
+	"io"
+	"io/ioutil"
+	"log"
+	"net/http"
+	"net/url"
+	"os"
+	//"os/exec"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
+
+	"encoding/base32"
+	"encoding/base64"
+	"encoding/binary"
+	"encoding/hex"
+	"encoding/json"
+	"encoding/xml"
+
+	"crypto/aes"
+	"crypto/cipher"
+	"crypto/hmac"
+	"crypto/md5"
+	"crypto/rand"
+	"crypto/sha1"
+
+	"github.com/bradfitz/slice"
+	qrcode "github.com/skip2/go-qrcode"
+
+	"github.com/frikky/kin-openapi/openapi2"
+	"github.com/frikky/kin-openapi/openapi2conv"
+	"github.com/frikky/kin-openapi/openapi3"
+
+	"github.com/satori/go.uuid"
+	"google.golang.org/appengine"
+)
 
 func HandleGetOrgs(resp http.ResponseWriter, request *http.Request) {
 	cors := HandleCors(resp, request)
@@ -9,7 +52,8 @@ func HandleGetOrgs(resp http.ResponseWriter, request *http.Request) {
 	user, err := HandleApiAuthentication(resp, request)
 	if err != nil {
 		log.Printf("[WARNING] Api authentication failed in get orgs: %s", err)
-        setResponse(resp, request, 401, `{"success": false}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
 		return
 	}
 
@@ -33,31 +77,37 @@ func HandleGetOrgs(resp http.ResponseWriter, request *http.Request) {
 		newjson, err := json.Marshal(orgs)
 		if err != nil {
 			log.Printf("[WARNING] Failed marshal in get orgs: %s", err)
-            setResponse(resp, request, 401, `{"success": false, "reason": "Failed unpacking"}`)
+			resp.WriteHeader(401)
+			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
 			return
 		}
 
 		//log.Printf("[AUDIT] User %s (%s) isn't global admin and can't list orgs. Returning list of local orgs.", user.Username, user.Id)
-        setResponse(resp, request, 200, newjson)
+		resp.WriteHeader(200)
+		resp.Write([]byte(newjson))
 		return
 	}
 
 	orgs, err := GetAllOrgs(ctx)
 	if err != nil || len(orgs) == 0 {
 		log.Printf("[WARNING] Failed getting orgs: %s", err)
-        setResponse(resp, request, 401, `{"success": false, "reason": "Can't get orgs"}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Can't get orgs"}`))
 		return
 	}
 
 	newjson, err := json.Marshal(orgs)
 	if err != nil {
 		log.Printf("[WARNING] Failed unmarshal in get orgs: %s", err)
-        setResponse(resp, request, 401, `{"success": false, "reason": "Failed unpacking"}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
 		return
 	}
 
-    setResponse(resp, request, 200, newjson)
+	resp.WriteHeader(200)
+	resp.Write(newjson)
 }
+
 
 func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
 	cors := HandleCors(resp, request)
@@ -70,7 +120,8 @@ func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
 	if location[1] == "api" {
 		if len(location) <= 4 {
 			log.Printf("Path too short: %d", len(location))
-            setResponse(resp, request, 401, `{"success": false}`)
+			resp.WriteHeader(401)
+			resp.Write([]byte(`{"success": false}`))
 			return
 		}
 
@@ -80,14 +131,16 @@ func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
 	user, err := HandleApiAuthentication(resp, request)
 	if err != nil {
 		log.Printf("[WARNING] Api authentication failed in get org: %s", err)
-        setResponse(resp, request, 401, `{"success": false}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
 		return
 	}
 
 	ctx := GetContext(request)
 	org, err := GetOrg(ctx, fileId)
 	if err != nil {
-        setResponse(resp, request, 401, `{"success": false, "reason": "Failed getting org details"}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Failed getting org details"}`))
 		return
 	}
 
@@ -107,7 +160,8 @@ func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
 
 	if !userFound {
 		log.Printf("[ERROR] User %s (%s) isn't a part of org %s (get)", user.Username, user.Id, org.Id)
-        setResponse(resp, request, 401, `{"success": false, "reason": "User doesn't have access to org"}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "User doesn't have access to org"}`))
 		return
 
 	}
@@ -181,9 +235,468 @@ func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
 	newjson, err := json.Marshal(org)
 	if err != nil {
 		log.Printf("[ERROR] Failed unmarshal of org %s (%s): %s", org.Name, org.Id, err)
-        setResponse(resp, request, 401, `{"success": false, "reason": "Failed unpacking"}`)
+		resp.WriteHeader(401)
+		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
 		return
 	}
 
-    setResponse(resp, request, 200, newjson)
+	resp.WriteHeader(200)
+	resp.Write(newjson)
+}
+
+func HandleCreateSubOrg(resp http.ResponseWriter, request *http.Request) {
+	cors := HandleCors(resp, request)
+	if cors {
+		return
+	}
+
+	user, err := HandleApiAuthentication(resp, request)
+	if err != nil {
+		log.Printf("[WARNING] Api authentication failed in creating sub org: %s", err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	if user.Role != "admin" {
+		log.Printf("[WARNING] Not admin: %s (%s).", user.Username, user.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Not admin"}`))
+		return
+	}
+
+	body, err := ioutil.ReadAll(request.Body)
+	if err != nil {
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
+		return
+	}
+
+	type ReturnData struct {
+		OrgId string `json:"org_id" datastore:"org_id"`
+		Name  string `json:"name" datastore:"name"`
+	}
+
+	var tmpData ReturnData
+	err = json.Unmarshal(body, &tmpData)
+	if err != nil {
+		log.Printf("Failed unmarshalling test: %s", err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	if len(tmpData.Name) < 3 {
+		log.Printf("[WARNING] Suborgname too short (min 3) %s", tmpData.Name)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Name must at least be 3 characters."}`))
+		return
+	}
+
+	var fileId string
+	location := strings.Split(request.URL.String(), "/")
+	if location[1] == "api" {
+		if len(location) <= 4 {
+			log.Printf("Path too short: %d", len(location))
+			resp.WriteHeader(401)
+			resp.Write([]byte(`{"success": false}`))
+			return
+		}
+
+		fileId = location[4]
+	}
+
+	if tmpData.OrgId != user.ActiveOrg.Id || fileId != user.ActiveOrg.Id {
+		log.Printf("[WARNING] User can't edit the org \"%s\"", tmpData.OrgId)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "No permission to edit this org"}`))
+		return
+	}
+
+	ctx := GetContext(request)
+	parentOrg, err := GetOrg(ctx, tmpData.OrgId)
+	if err != nil {
+		log.Printf("[WARNING] Organization %s doesn't exist: %s", tmpData.OrgId, err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	if len(parentOrg.ManagerOrgs) > 0 {
+		log.Printf("[WARNING] Organization %s can't have suborgs, as it's as suborg", tmpData.OrgId)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Can't make suborg of suborg. Switch to a parent org to make another."}`))
+		return
+	}
+
+	if project.Environment == "cloud" {
+		if !parentOrg.SyncFeatures.MultiTenant.Active {
+			log.Printf("[WARNING] Org %s is not allowed to make a sub-organization: %s", tmpData.OrgId, err)
+			resp.WriteHeader(401)
+			resp.Write([]byte(`{"success": false, "reason": "Sub-organizations require an active subscription with access to multi-tenancy. Contact support to try it out."}`))
+			return
+		}
+
+		/*
+			if parentOrg.SyncUsage.MultiTenant.Counter >= parentOrg.SyncFeatures.MultiTenant.Limit || len(parentOrg.ChildOrgs) > int(parentOrg.SyncFeatures.MultiTenant.Limit) {
+				log.Printf("[WARNING] Org %s is not allowed to make ANOTHER sub-organization. Limit reached!: %s", tmpData.OrgId, err)
+				resp.WriteHeader(401)
+				resp.Write([]byte(`{"success": false, "reason": "Your limit of sub-organizations has been reached. Contact support to increase."}`))
+				return
+			}
+		*/
+
+		parentOrg.SyncUsage.MultiTenant.Counter += 1
+		log.Printf("[DEBUG] Allowing suborg for %s because they have %d vs %d limit", parentOrg.Id, len(parentOrg.ChildOrgs), parentOrg.SyncFeatures.MultiTenant.Limit)
+	}
+
+	orgId := uuid.NewV4().String()
+	newOrg := Org{
+		Name:        tmpData.Name,
+		Description: fmt.Sprintf("Sub-org by user %s in parent-org %s", user.Username, parentOrg.Name),
+		Image:       parentOrg.Image,
+		Id:          orgId,
+		Org:         tmpData.Name,
+		Users: []User{
+			user,
+		},
+		Roles:     []string{"admin", "user"},
+		CloudSync: false,
+		ManagerOrgs: []OrgMini{
+			OrgMini{
+				Id:   tmpData.OrgId,
+				Name: parentOrg.Name,
+			},
+		},
+		CloudSyncActive: parentOrg.CloudSyncActive,
+		CreatorOrg:      tmpData.OrgId,
+	}
+	//SyncFeatures:    parentOrg.SyncFeatures,
+
+	parentOrg.ChildOrgs = append(parentOrg.ChildOrgs, OrgMini{
+		Name: tmpData.Name,
+		Id:   orgId,
+	})
+
+	err = SetOrg(ctx, newOrg, newOrg.Id)
+	if err != nil {
+		log.Printf("[WARNING] Failed setting new org %s: %s", newOrg.Id, err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	err = SetOrg(ctx, *parentOrg, parentOrg.Id)
+	if err != nil {
+		log.Printf("[WARNING] Failed updating parent org %s: %s", newOrg.Id, err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	// Update all admins to have access to this suborg
+	for _, loopUser := range parentOrg.Users {
+		if loopUser.Role != "admin" {
+			continue
+		}
+
+		if loopUser.Id == user.Id {
+			continue
+		}
+
+		foundUser, err := GetUser(ctx, loopUser.Id)
+		if err != nil {
+			log.Printf("[WARNING] User with Identifier %s doesn't exist: %s (update admins - create)", loopUser.Id, err)
+			continue
+		}
+
+		foundUser.Orgs = append(foundUser.Orgs, newOrg.Id)
+		err = SetUser(ctx, foundUser, false)
+		if err != nil {
+			log.Printf("[WARNING] Failed updating user when setting creating suborg (update admins - update): %s ", err)
+			continue
+		}
+	}
+
+	user.Orgs = append(user.Orgs, newOrg.Id)
+	log.Printf("[INFO] Usr: %#v (%d)", user.Orgs, len(user.Orgs))
+	err = SetUser(ctx, &user, false)
+	if err != nil {
+		log.Printf("[WARNING] Failed updating user when setting creating suborg: %s", err)
+		resp.WriteHeader(500)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	/*
+		if project.Environment != "cloud" {
+			log.Printf("[DEBUG] Starting cloud schedule for org %s (%s)", newOrg.Name, newOrg.Id)
+			interval := 15
+			log.Printf("[DEBUG] Should start schedule for org %s", newOrg.Name)
+			job := func() {
+				err := remoteOrgJobHandler(newOrg, interval)
+				if err != nil {
+					log.Printf("[ERROR] Failed request with remote org setup (3): %s", err)
+				}
+			}
+
+			jobret, err := newscheduler.Every(int(interval)).Seconds().NotImmediately().Run(job)
+			if err != nil {
+				log.Printf("[CRITICAL] Failed to schedule new org: %s", err)
+			}
+		}
+	*/
+
+	log.Printf("[INFO] User %s SUCCESSFULLY ADDED child org %s (%s) for parent %s (%s)", user.Username, newOrg.Name, newOrg.Id, parentOrg.Name, parentOrg.Id)
+	resp.WriteHeader(200)
+	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully updated org"}`)))
+
+}
+
+func HandleEditOrg(resp http.ResponseWriter, request *http.Request) {
+	cors := HandleCors(resp, request)
+	if cors {
+		return
+	}
+
+	user, err := HandleApiAuthentication(resp, request)
+	if err != nil {
+		log.Printf("[WARNING] Api authentication failed in edit org: %s", err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	if user.Role != "admin" {
+		log.Printf("[WARNING] Not admin: %s (%s).", user.Username, user.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Not admin"}`))
+		return
+	}
+
+	body, err := ioutil.ReadAll(request.Body)
+	if err != nil {
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
+		return
+	}
+
+	type ReturnData struct {
+		Tutorial    string    `json:"tutorial" datastore:"tutorial"`
+		Name        string    `json:"name" datastore:"name"`
+		Image       string    `json:"image" datastore:"image"`
+		CompanyType string    `json:"company_type" datastore:"company_type"`
+		Description string    `json:"description" datastore:"description"`
+		OrgId       string    `json:"org_id" datastore:"org_id"`
+		Priority    string    `json:"priority" datastore:"priority"`
+		Defaults    Defaults  `json:"defaults" datastore:"defaults"`
+		SSOConfig   SSOConfig `json:"sso_config" datastore:"sso_config"`
+	}
+
+	var tmpData ReturnData
+	err = json.Unmarshal(body, &tmpData)
+	if err != nil {
+		log.Printf("Failed unmarshalling test: %s", err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+	//log.Printf("SSO: %#v", tmpData.SSOConfig)
+
+	var fileId string
+	location := strings.Split(request.URL.String(), "/")
+	if location[1] == "api" {
+		if len(location) <= 4 {
+			log.Printf("Path too short: %d", len(location))
+			resp.WriteHeader(401)
+			resp.Write([]byte(`{"success": false}`))
+			return
+		}
+
+		fileId = location[4]
+	}
+
+	if tmpData.OrgId != user.ActiveOrg.Id || fileId != user.ActiveOrg.Id {
+		log.Printf("[WARNING] User can't edit org %#v (active: %#v)", fileId, user.ActiveOrg.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false, "No permission to edit this org"}`))
+		return
+	}
+
+	ctx := GetContext(request)
+	org, err := GetOrg(ctx, tmpData.OrgId)
+	if err != nil {
+		log.Printf("[WARNING] Organization doesn't exist: %s", err)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	admin := false
+	userFound := false
+	for _, inneruser := range org.Users {
+		if inneruser.Id == user.Id {
+			userFound = true
+			if inneruser.Role == "admin" {
+				admin = true
+			}
+
+			break
+		}
+	}
+
+	if !userFound {
+		log.Printf("User %s doesn't exist in organization for edit %s", user.Id, org.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	if !admin {
+		log.Printf("User %s doesn't have edit rights to %s", user.Id, org.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	sendOrgUpdaterHook := false
+	if len(tmpData.Image) > 0 {
+		org.Image = tmpData.Image
+	}
+
+	if len(tmpData.Name) > 0 {
+		org.Name = tmpData.Name
+	}
+
+	if len(tmpData.Description) > 0 {
+		org.Description = tmpData.Description
+	}
+
+	if len(tmpData.Defaults.AppDownloadRepo) > 0 || len(tmpData.Defaults.AppDownloadBranch) > 0 || len(tmpData.Defaults.WorkflowDownloadRepo) > 0 || len(tmpData.Defaults.WorkflowDownloadBranch) > 0 || len(tmpData.Defaults.NotificationWorkflow) > 0 {
+		org.Defaults = tmpData.Defaults
+	}
+
+	if len(tmpData.CompanyType) > 0 {
+		org.CompanyType = tmpData.CompanyType
+
+		if len(org.CompanyType) == 0 {
+			sendOrgUpdaterHook = true
+		}
+	}
+
+	if len(tmpData.Tutorial) > 0 {
+		if tmpData.Tutorial == "welcome" {
+			sendOrgUpdaterHook = true
+		}
+	}
+
+	if len(tmpData.Priority) > 0 {
+		if len(org.MainPriority) == 0 {
+			org.MainPriority = tmpData.Priority
+			sendOrgUpdaterHook = true
+		}
+
+		found := false
+		for _, prio := range org.Priorities {
+			if prio.Name == tmpData.Priority {
+				found = true
+			}
+		}
+
+		if !found {
+			org.Priorities = append(org.Priorities, Priority{
+				Name:        tmpData.Priority,
+				Description: fmt.Sprintf("Priority %s decided by user.", tmpData.Priority),
+				Type:        "usecases",
+				Active:      true,
+				URL:         fmt.Sprintf("/usecases"),
+			})
+		}
+	}
+
+	//if len(tmpData.SSOConfig) > 0 {
+	if len(tmpData.SSOConfig.SSOEntrypoint) > 0 || len(tmpData.SSOConfig.OpenIdClientId) > 0 || len(tmpData.SSOConfig.OpenIdClientSecret) > 0 || len(tmpData.SSOConfig.OpenIdAuthorization) > 0 || len(tmpData.SSOConfig.OpenIdToken) > 0 {
+		org.SSOConfig = tmpData.SSOConfig
+	}
+
+	if len(tmpData.SSOConfig.SSOCertificate) > 0 {
+		savedCert := fixCertificate(tmpData.SSOConfig.SSOCertificate)
+
+		log.Printf("[INFO] Stripped down cert from %d to %d", len(tmpData.SSOConfig.SSOCertificate), len(savedCert))
+
+		org.SSOConfig.SSOCertificate = savedCert
+	}
+
+	if len(org.Defaults.NotificationWorkflow) > 0 && len(org.Defaults.NotificationWorkflow) != 36 {
+		log.Printf("[WARNING] Notification Workflow ID %s is not valid.", org.Defaults.NotificationWorkflow)
+	}
+
+	// Built a system around this now, which checks for the actual org. Only works onprem so far.
+	// if requestdata.Environment == "cloud" && project.Environment != "cloud" {
+	//if project.Environment != "cloud" && len(org.SSOConfig.SSOEntrypoint) > 0 && len(org.ManagerOrgs) == 0 {
+	//	//log.Printf("[INFO] Should set SSO entrypoint to %s", org.SSOConfig.SSOEntrypoint)
+	//	SSOUrl = org.SSOConfig.SSOEntrypoint
+	//}
+
+	//log.Printf("Org: %#v", org)
+	err = SetOrg(ctx, *org, org.Id)
+	if err != nil {
+		log.Printf("User %s doesn't have edit rights to %s", user.Id, org.Id)
+		resp.WriteHeader(401)
+		resp.Write([]byte(`{"success": false}`))
+		return
+	}
+
+	// Only send for cloud
+	if sendOrgUpdaterHook && project.Environment == "cloud" {
+		signupWebhook := os.Getenv("WEBSITE_ORG_WEBHOOK")
+		if strings.HasPrefix(signupWebhook, "http") {
+			curIndex := -1
+			for orgIndex, orguser := range org.Users {
+				if orguser.Id == user.Id {
+					curIndex = orgIndex
+				}
+			}
+
+			if curIndex >= 0 {
+				user.Password = ""
+				user.Session = ""
+				user.ApiKey = ""
+				user.LoginInfo = []LoginInfo{}
+				user.PrivateApps = []WorkflowApp{}
+
+				org.Users[curIndex] = user
+			}
+
+			mappedData, err := json.Marshal(org)
+			if err != nil {
+				log.Printf("[WARNING] Marshal error for org sending: %s", err)
+			} else {
+				req, err := http.NewRequest(
+					"POST",
+					signupWebhook,
+					bytes.NewBuffer(mappedData),
+				)
+
+				client := &http.Client{
+					Timeout: 3 * time.Second,
+				}
+
+				req.Header.Add("Content-Type", "application/json")
+				res, err := client.Do(req)
+				if err != nil {
+					log.Printf("[ERROR] Failed request to signup webhook FOR ORG (2): %s", err)
+				} else {
+					log.Printf("[INFO] Successfully ran org priority webhook")
+				}
+
+				_ = res
+			}
+		}
+	}
+
+	log.Printf("[INFO] Successfully updated org %s (%s)", org.Name, org.Id)
+	resp.WriteHeader(200)
+	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully updated org"}`)))
+
 }
