+++ shared.go
@@ -448,114 +448,6 @@ func randStr(strSize int, randType string) string {
 	return string(bytes)
 }
 
-func HandleSet2fa(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[ERROR] Api authentication failed in get 2fa: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("[ERROR] Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("Error with body read: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	type parsedValue struct {
-		Code   string `json:"code"`
-		UserId string `json:"user_id"`
-	}
-
-	var tmpBody parsedValue
-	err = json.Unmarshal(body, &tmpBody)
-	if err != nil {
-		log.Printf("[WARNING] Error with unmarshal tmpBody in verify 2fa: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if len(tmpBody.Code) != 6 {
-		log.Printf("[WARNING] Length of code isn't 6: %s", tmpBody.Code)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Length of code must be 6"}`)))
-		return
-	}
-
-	// FIXME: Everything should match?
-	// || user.Id != tmpBody.UserId
-	if user.Id != fileId {
-		log.Printf("[WARNING] Bad ID: %s vs %s", user.Id, fileId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Can only set 2fa for your own user. Pass field user_id in JSON."}`)))
-		return
-	}
-
-	ctx := GetContext(request)
-	foundUser, err := GetUser(ctx, user.Id)
-	if err != nil {
-		log.Printf("[ERROR] Can't find user %s (set 2fa): %s", user.Id, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed getting your user."}`)))
-		return
-	}
-
-	//https://www.gojek.io/blog/a-diy-two-factor-authenticator-in-golang
-	interval := time.Now().Unix() / 30
-	HOTP, err := getHOTPToken(foundUser.MFA.PreviousCode, interval)
-	if err != nil {
-		log.Printf("[ERROR] Failed generating a HOTP token: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if HOTP != tmpBody.Code {
-		log.Printf("[DEBUG] Bad code sent for user %s (%s). Sent: %s, Want: %s", user.Username, user.Id, tmpBody.Code, HOTP)
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false, "reason": "Wrong code. Try again"}`))
-		return
-	}
-
-	foundUser.MFA.Active = true
-	foundUser.MFA.ActiveCode = foundUser.MFA.PreviousCode
-	foundUser.MFA.PreviousCode = ""
-	err = SetUser(ctx, foundUser, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed SETTING MFA for user %s (%s): %s", foundUser.Username, foundUser.Id, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed updating your user. Please try again."}`))
-		return
-	}
-
-	log.Printf("[DEBUG] Successfully enabled 2FA for user %s (%s)", foundUser.Username, foundUser.Id)
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(`{"success": true, "reason": "Correct code. MFA is now required for this user."}`))
-}
-
 func getHOTPToken(secret string, interval int64) (string, error) {
 
 	//Converts secret to base32 Encoding. Base32 encoding desires a 32-character
@@ -611,644 +503,6 @@ func getHOTPToken(secret string, interval int64) (string, error) {
 	}
 
 	return otp, nil
-}
-
-func HandleGet2fa(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[ERROR] Api authentication failed in get 2fa: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("[ERROR] Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if user.Id != fileId {
-		log.Printf("[WARNING] Bad ID: %s vs %s", user.Id, fileId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Can only set 2fa for your own user"}`)))
-		return
-	}
-
-	// https://socketloop.com/tutorials/golang-generate-qr-codes-for-google-authenticator-app-and-fix-cannot-interpret-qr-code-error
-
-	// generate a random string - preferably 6 or 8 characters
-	randomStr := randStr(8, "alphanum")
-	//fmt.Println(randomStr)
-
-	// For Google Authenticator purpose
-	// for more details see
-	// https://github.com/google/google-authenticator/wiki/Key-Uri-Format
-	secret := base32.StdEncoding.EncodeToString([]byte(randomStr))
-
-	// authentication link. Remember to replace SocketLoop with yours.
-	// for more details see
-	// https://github.com/google/google-authenticator/wiki/Key-Uri-Format
-	authLink := fmt.Sprintf("otpauth://totp/%s?secret=%s&issuer=Shuffle", user.Username, secret)
-	png, err := qrcode.Encode(authLink, qrcode.Medium, 256)
-	if err != nil {
-		log.Printf("[ERROR] Failed PNG encoding: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed image encoding"}`)))
-		return
-	}
-
-	dataURI := fmt.Sprintf("data:image/png;base64,%s", base64.StdEncoding.EncodeToString([]byte(png)))
-	newres := ResultChecker{
-		Success: true,
-		Reason:  dataURI,
-		Extra:   strings.ReplaceAll(secret, "=", "A"),
-	}
-
-	newjson, err := json.Marshal(newres)
-	if err != nil {
-		log.Printf("[ERROR] Failed marshal in get OTP: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking data"}`)))
-		return
-	}
-
-	ctx := GetContext(request)
-	//user.MFA.PreviousCode = authLink
-	user.MFA.PreviousCode = secret
-	err = SetUser(ctx, &user, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating MFA for user: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed updating your user"}`))
-		return
-	}
-
-	log.Printf("[DEBUG] Sent new MFA update for user %s (%s)", user.Username, user.Id)
-	//log.Printf("%s", newjson)
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(newjson))
-}
-
-func HandleGetOrgs(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in get orgs: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	if user.Role != "global_admin" {
-		orgs := []OrgMini{}
-		for _, item := range user.Orgs {
-			// FIXM: Should return normal orgs, but hidden if the user isn't admin
-			org, err := GetOrg(ctx, item)
-			if err == nil {
-				orgs = append(orgs, OrgMini{
-					Id:         org.Id,
-					Name:       org.Name,
-					CreatorOrg: org.CreatorOrg,
-					Image:      org.Image,
-				})
-				// Role:       "admin",
-			}
-		}
-
-		newjson, err := json.Marshal(orgs)
-		if err != nil {
-			log.Printf("[WARNING] Failed marshal in get orgs: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
-			return
-		}
-
-		//log.Printf("[AUDIT] User %s (%s) isn't global admin and can't list orgs. Returning list of local orgs.", user.Username, user.Id)
-		resp.WriteHeader(200)
-		resp.Write([]byte(newjson))
-		return
-	}
-
-	orgs, err := GetAllOrgs(ctx)
-	if err != nil || len(orgs) == 0 {
-		log.Printf("[WARNING] Failed getting orgs: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Can't get orgs"}`))
-		return
-	}
-
-	newjson, err := json.Marshal(orgs)
-	if err != nil {
-		log.Printf("[WARNING] Failed unmarshal in get orgs: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write(newjson)
-}
-
-func HandleGetOrg(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in get org: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	org, err := GetOrg(ctx, fileId)
-	if err != nil {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed getting org details"}`))
-		return
-	}
-
-	admin := false
-	userFound := false
-	for _, inneruser := range org.Users {
-		if inneruser.Id == user.Id {
-			userFound = true
-
-			if inneruser.Role == "admin" {
-				admin = true
-			}
-
-			break
-		}
-	}
-
-	if !userFound {
-		log.Printf("[ERROR] User %s (%s) isn't a part of org %s (get)", user.Username, user.Id, org.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "User doesn't have access to org"}`))
-		return
-
-	}
-
-	if !admin {
-		org.Defaults = Defaults{}
-		org.SSOConfig = SSOConfig{}
-		org.Subscriptions = []PaymentSubscription{}
-		org.ManagerOrgs = []OrgMini{}
-		org.ChildOrgs = []OrgMini{}
-		org.Invites = []string{}
-	} else {
-		org.SyncFeatures.AppExecutions.Description = "The amount of Apps within Workflows you can run per month. This limit can be exceeded when running workflows without a trigger (manual execution)."
-		org.SyncFeatures.WorkflowExecutions.Description = "N/A. See App Executions"
-		org.SyncFeatures.Webhook.Description = "Webhooks are Triggers that take an HTTP input to start a workflow. Read docs for more."
-		org.SyncFeatures.Schedules.Description = "Schedules are Triggers that run on an interval defined by you. Read docs for more."
-		org.SyncFeatures.MultiEnv.Description = "Multiple Environments are used to run automation in different physical locations. Change from /admin?tab=environments"
-		org.SyncFeatures.MultiTenant.Description = "Multiple Tenants can be used to segregate information for each MSSP Customer. Change from /admin?tab=suborgs"
-		//org.SyncFeatures.MultiTenant.Description = "Multiple Tenants can be used to segregate information for each MSSP Customer. Change from /admin?tab=suborgs"
-
-		//log.Printf("LIMIT: %#v", org.SyncFeatures.AppExecutions.Limit)
-		orgChanged := false
-		if org.SyncFeatures.AppExecutions.Limit == 0 || org.SyncFeatures.AppExecutions.Limit == 1500 {
-			org.SyncFeatures.AppExecutions.Limit = 5000
-			orgChanged = true
-		}
-
-		if org.SyncFeatures.SendMail.Limit == 0 {
-			org.SyncFeatures.SendMail.Limit = 100
-			orgChanged = true
-		}
-
-		if org.SyncFeatures.SendSms.Limit == 0 {
-			org.SyncFeatures.SendSms.Limit = 30
-			orgChanged = true
-		}
-
-		org.SyncFeatures.EmailTrigger.Limit = 0
-		if org.SyncFeatures.MultiEnv.Limit == 0 {
-			org.SyncFeatures.MultiEnv.Limit = 1
-			orgChanged = true
-		}
-
-		org.SyncFeatures.EmailTrigger.Limit = 0
-
-		if orgChanged {
-			log.Printf("[DEBUG] Org features for %s (%s) changed. Updating.", org.Name, org.Id)
-			err = SetOrg(ctx, *org, org.Id)
-			if err != nil {
-				log.Printf("[WARNING] Failed updating org during org loading")
-			}
-		}
-
-		info, err := GetOrgStatistics(ctx, fileId)
-		if err == nil {
-			org.SyncFeatures.AppExecutions.Usage = info.MonthlyAppExecutions
-		}
-
-		org.SyncFeatures.MultiTenant.Usage = int64(len(org.ChildOrgs) + 1)
-		envs, err := GetEnvironments(ctx, fileId)
-		if err == nil {
-			//log.Printf("Envs: %#v", len(envs))
-			org.SyncFeatures.MultiEnv.Usage = int64(len(envs))
-		}
-	}
-
-	org.Users = []User{}
-	org.SyncConfig.Apikey = ""
-	org.SyncConfig.Source = ""
-
-	newjson, err := json.Marshal(org)
-	if err != nil {
-		log.Printf("[ERROR] Failed unmarshal of org %s (%s): %s", org.Name, org.Id, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking"}`)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write(newjson)
-}
-
-func HandleLogout(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	ctx := GetContext(request)
-
-	runReturn := false
-	userInfo, usererr := HandleApiAuthentication(resp, request)
-	log.Printf("[AUDIT] Logging out user %s (%s)", userInfo.Username, userInfo.Id)
-
-	if project.Environment == "cloud" {
-		// Checking if it's a special region. All user-specific requests should
-		// go through shuffler.io and not subdomains
-		gceProject := os.Getenv("SHUFFLE_GCEPROJECT")
-		if gceProject != "shuffler" && len(gceProject) > 0 {
-			log.Printf("[DEBUG] Redirecting LOGOUT request to main site handler (shuffler.io)")
-
-			RedirectUserRequest(resp, request)
-			// FIXME: Allow superfluous cleanups?
-			// Point is: should it continue running the logout to
-			// ensure cookies are cleared?
-			// Keeping it for now to ensure cleanup.
-		}
-	}
-
-	c, err := request.Cookie("session_token")
-	if err != nil {
-		c, err = request.Cookie("__session")
-	}
-
-	if err == nil {
-		newCookie := &http.Cookie{
-			Name:    "session_token",
-			Value:   c.Value,
-			Expires: time.Now().Add(-100 * time.Hour),
-			MaxAge:  -1,
-		}
-		if project.Environment == "cloud" {
-			newCookie.Domain = ".shuffler.io"
-			newCookie.Secure = true
-			newCookie.HttpOnly = true
-		}
-
-		http.SetCookie(resp, newCookie)
-
-		newCookie.Name = "__session"
-		http.SetCookie(resp, newCookie)
-
-	} else {
-		newCookie := &http.Cookie{
-			Name:    "session_token",
-			Value:   "",
-			Expires: time.Now().Add(-100 * time.Hour),
-			MaxAge:  -1,
-		}
-
-		if project.Environment == "cloud" {
-			newCookie.Domain = ".shuffler.io"
-			newCookie.Secure = true
-			newCookie.HttpOnly = true
-		}
-
-		http.SetCookie(resp, newCookie)
-
-		newCookie.Name = "__session"
-		http.SetCookie(resp, newCookie)
-	}
-
-	if runReturn == true {
-		DeleteCache(ctx, fmt.Sprintf("user_%s", strings.ToLower(userInfo.Username)))
-		DeleteCache(ctx, fmt.Sprintf("session_%s", userInfo.Session))
-		DeleteCache(ctx, userInfo.Session)
-
-		log.Printf("[INFO] Returning from logout request after cache cleanup")
-
-		return
-	}
-
-	if usererr != nil {
-		log.Printf("[WARNING] Api authentication failed in handleLogout: %s", err)
-		resp.WriteHeader(200)
-		resp.Write([]byte(`{"success": true, "reason": "Not logged in"}`))
-		return
-	}
-
-	DeleteCache(ctx, fmt.Sprintf("user_%s", strings.ToLower(userInfo.Username)))
-	DeleteCache(ctx, fmt.Sprintf("session_%s", userInfo.Session))
-	DeleteCache(ctx, userInfo.Session)
-
-	userInfo.Session = ""
-	err = SetUser(ctx, &userInfo, true)
-	if err != nil {
-		log.Printf("Failed updating user: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed updating apikey"}`))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(`{"success": false, "reason": "Successfully logged out"}`))
-}
-
-func HandleApiAuthentication(resp http.ResponseWriter, request *http.Request) (User, error) {
-	apikey := request.Header.Get("Authorization")
-
-	user := &User{}
-	if len(apikey) > 0 {
-		if !strings.HasPrefix(apikey, "Bearer ") {
-
-			//location := strings.Split(request.URL.String(), "/")
-			if !strings.Contains(request.URL.String(), "/execute") {
-				log.Printf("[WARNING] Apikey doesn't start with bearer")
-			}
-
-			return User{}, errors.New("No bearer token for authorization header")
-		}
-
-		apikeyCheck := strings.Split(apikey, " ")
-		if len(apikeyCheck) != 2 {
-			log.Printf("[WARNING] Invalid format for apikey: %s", apikeyCheck)
-			return User{}, errors.New("Invalid format for apikey")
-		}
-
-		if len(apikeyCheck[1]) < 36 {
-			return User{}, errors.New("Apikey must be at least 36 characters long (UUID)")
-		}
-
-		// This is annoying af
-		newApikey := apikeyCheck[1]
-		if len(newApikey) > 249 {
-			newApikey = newApikey[0:248]
-		}
-
-		ctx := GetContext(request)
-		cache, err := GetCache(ctx, newApikey)
-		if err == nil {
-			cacheData := []byte(cache.([]uint8))
-			err = json.Unmarshal(cacheData, &user)
-			if err == nil {
-				//log.Printf("[WARNING] Got user from cache: %s", err)
-
-				if len(user.Id) == 0 && len(user.Username) == 0 {
-					return User{}, errors.New(fmt.Sprintf("Couldn't find user"))
-				}
-
-				user.SessionLogin = false
-				return *user, nil
-			}
-		} else {
-			//log.Printf("[WARNING] Error getting authentication cache for %s: %v", newApikey, err)
-		}
-
-		// Make specific check for just service user?
-		// Get the user based on APIkey here
-		//log.Println(apikeyCheck[1])
-		userdata, err := GetApikey(ctx, apikeyCheck[1])
-		if err != nil {
-			log.Printf("[WARNING] Apikey %s doesn't exist: %s", apikey, err)
-			return User{}, err
-		}
-
-		if len(userdata.Id) == 0 && len(userdata.Username) == 0 {
-			log.Printf("[WARNING] Apikey %s doesn't exist or the user doesn't have an ID/Username", apikey)
-			return User{}, errors.New("Couldn't find the user")
-		}
-
-		// Caching both bad and good apikeys :)
-		b, err := json.Marshal(userdata)
-		if err != nil {
-			log.Printf("[WARNING] Failed marshalling: %s", err)
-			return User{}, err
-		}
-
-		user.SessionLogin = false
-		err = SetCache(ctx, newApikey, b)
-		if err != nil {
-			log.Printf("[WARNING] Failed setting cache for apikey: %s", err)
-		}
-
-		return userdata, nil
-	}
-
-	// One time API keys
-	authorizationArr, ok := request.URL.Query()["authorization"]
-	ctx := GetContext(request)
-	if ok {
-		authorization := ""
-		if len(authorizationArr) > 0 {
-			authorization = authorizationArr[0]
-		}
-		_ = authorization
-		log.Printf("[ERROR] WHAT ARE ONE TIME KEYS USED FOR?")
-	}
-
-	c, err := request.Cookie("session_token")
-	// Compatibility issues
-	if err != nil {
-		c, err = request.Cookie("__session")
-	}
-
-	if err == nil {
-		sessionToken := c.Value
-		user, err := GetSessionNew(ctx, sessionToken)
-		if err != nil {
-			log.Printf("[DEBUG] No valid session token for ID %s. Setting cookie to expire.", sessionToken)
-
-			newCookie := &http.Cookie{
-				Name:    "session_token",
-				Value:   sessionToken,
-				Expires: time.Now().Add(-100 * time.Hour),
-				MaxAge:  -1,
-			}
-
-			if project.Environment == "cloud" {
-				newCookie.Domain = ".shuffler.io"
-				newCookie.Secure = true
-				newCookie.HttpOnly = true
-			}
-
-			http.SetCookie(resp, newCookie)
-
-			newCookie.Name = "__session"
-			http.SetCookie(resp, newCookie)
-
-			return User{}, err
-		}
-
-		//user, err := GetUser(ctx, session.UserId)
-		//if err != nil {
-		//	log.Printf("[INFO] User with Identifier %s doesn't exist: %s", session.UserId, err)
-		//	http.SetCookie(resp, &http.Cookie{
-		//		Name:    "session_token",
-		//		Value:   sessionToken,
-		//		Expires: time.Now().Add(-100 * time.Hour),
-		//		MaxAge:  -1,
-		//	})
-
-		//	return User{}, err
-		//}
-
-		if len(user.Id) == 0 && len(user.Username) == 0 {
-
-			newCookie := &http.Cookie{
-				Name:    "session_token",
-				Value:   sessionToken,
-				Expires: time.Now().Add(-100 * time.Hour),
-				MaxAge:  -1,
-			}
-
-			if project.Environment == "cloud" {
-				newCookie.Domain = ".shuffler.io"
-				newCookie.Secure = true
-				newCookie.HttpOnly = true
-			}
-
-			http.SetCookie(resp, newCookie)
-
-			newCookie.Name = "__session"
-			http.SetCookie(resp, newCookie)
-
-			return User{}, errors.New(fmt.Sprintf("Couldn't find user"))
-		}
-
-		// We're using the session to find the user anyway, which is NOT user controlled
-		// This means that this is redundant, but MAY allow users
-		// to have access past session timeouts
-		//if user.Session != sessionToken {
-		//	return User{}, errors.New("[WARNING] Wrong session token")
-		//}
-
-		user.SessionLogin = true
-
-		// Means session exists, but
-		return user, nil
-	}
-
-	// Key = apikey
-	return User{}, errors.New("Missing authentication")
-}
-
-func GetOpenapi(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	// Just here to verify that the user is logged in
-	_, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in validate swagger: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	location := strings.Split(request.URL.String(), "/")
-	var id string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("Missing parts of API in request!")
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		id = location[4]
-	}
-
-	/*
-		if len(id) != 32 {
-			log.Printf("Missing parts of API in request!")
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-	*/
-	//_, err = GetApp(ctx, id)
-	//if err == nil {
-	//	log.Println("You're supposed to be able to continue now.")
-	//}
-
-	// FIXME - FIX AUTH WITH APP
-	ctx := GetContext(request)
-	parsedApi, err := GetOpenApiDatastore(ctx, id)
-	if err != nil {
-		log.Printf("[ERROR] Failed getting OpenAPI: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	log.Printf("[INFO] API LENGTH GET: %d, ID: %s", len(parsedApi.Body), id)
-
-	parsedApi.Success = true
-	data, err := json.Marshal(parsedApi)
-	if err != nil {
-		log.Printf("[ERROR] Failed unmarshaling OpenAPI: %s", err)
-		resp.WriteHeader(422)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed marshalling parsed swagger: %s"}`, err)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write(data)
 }
 
 func GetResult(ctx context.Context, workflowExecution WorkflowExecution, id string) (WorkflowExecution, ActionResult) {
@@ -1485,70 +739,6 @@ func SetAuthenticationConfig(resp http.ResponseWriter, request *http.Request) {
 	//log.Printf("Should set %s
 }
 
-func HandleGetSchedules(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in get schedules: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Role != "admin" {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Admin required"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	schedules, err := GetAllSchedules(ctx, user.ActiveOrg.Id)
-	if err != nil {
-		log.Printf("[WARNING] Failed getting schedules: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Couldn't get schedules"}`))
-		return
-	}
-
-	newjson, err := json.Marshal(schedules)
-	if err != nil {
-		log.Printf("Failed unmarshal: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unpacking environments"}`)))
-		return
-	}
-
-	//log.Printf("Existing environments: %s", string(newjson))
-
-	resp.WriteHeader(200)
-	resp.Write(newjson)
-}
-
-func GenerateApikey(ctx context.Context, userInfo User) (User, error) {
-	// Generate UUID
-	// Set uuid to apikey in backend (update)
-	userInfo.ApiKey = uuid.NewV4().String()
-	err := SetApikey(ctx, userInfo)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating apikey: %s", err)
-		return userInfo, err
-	}
-
-	// Updating user
-	log.Printf("[INFO] Adding apikey to user %s", userInfo.Username)
-	err = SetUser(ctx, &userInfo, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating users' apikey: %s", err)
-		return userInfo, err
-	}
-
-	return userInfo, nil
-}
-
 func HandleCategoryIncrease(categories Categories, action Action, workflowapps []WorkflowApp) Categories {
 	if action.Category == "" {
 		appName := action.AppName
@@ -1593,280 +783,6 @@ func HandleCategoryIncrease(categories Categories, action Action, workflowapps [
 
 	return categories
 }
-
-func HandleApiGeneration(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	if project.Environment == "cloud" {
-		// Checking if it's a special region. All user-specific requests should
-		// go through shuffler.io and not subdomains
-		gceProject := os.Getenv("SHUFFLE_GCEPROJECT")
-		if gceProject != "shuffler" && len(gceProject) > 0 {
-			log.Printf("[DEBUG] Redirecting API GEN request to main site handler (shuffler.io)")
-			RedirectUserRequest(resp, request)
-			return
-		}
-	}
-
-	userInfo, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in apigen: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	//log.Printf("IN APIKEY GENERATION")
-	ctx := GetContext(request)
-	if request.Method == "GET" {
-		newUserInfo, err := GenerateApikey(ctx, userInfo)
-		if err != nil {
-			log.Printf("[WARNING] Failed to generate apikey for user %s: %s", userInfo.Username, err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false, "reason": ""}`))
-			return
-		}
-
-		userInfo = newUserInfo
-		log.Printf("[INFO] Updated apikey for user %s", userInfo.Username)
-	} else if request.Method == "POST" {
-		if request.Body == nil {
-			resp.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		body, err := ioutil.ReadAll(request.Body)
-		if err != nil {
-			log.Println("Failed reading body")
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Missing field: user_id"}`)))
-			return
-		}
-
-		type userId struct {
-			UserId string `json:"user_id"`
-		}
-
-		var t userId
-		err = json.Unmarshal(body, &t)
-		if err != nil {
-			log.Printf("Failed unmarshaling userId: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed unmarshaling. Missing field: user_id"}`)))
-			return
-		}
-
-		log.Printf("[INFO] Handling post for APIKEY gen FROM user %s. Userchange: %s!", userInfo.Username, t.UserId)
-
-		if userInfo.Role != "admin" {
-			log.Printf("[AUDIT] %s tried and failed to change apikey for %s (2)", userInfo.Username, t.UserId)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "You need to be admin to change others' apikey"}`)))
-			return
-		}
-
-		foundUser, err := GetUser(ctx, t.UserId)
-		if err != nil {
-			log.Printf("[INFO] Can't find user %s (apikey gen): %s", t.UserId, err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false}`)))
-			return
-		}
-
-		// FIXME: May not be good due to different roles in different organizations.
-		if foundUser.Role == "admin" {
-			log.Printf("[AUDIT] %s tried and failed to change apikey for %s. Skipping because users' role is admin", userInfo.Username, t.UserId)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Can't change the apikey of another admin"}`)))
-			return
-		}
-
-		newUserInfo, err := GenerateApikey(ctx, *foundUser)
-		if err != nil {
-			log.Printf("Failed to generate apikey for user %s: %s", foundUser.Username, err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "%s"}`, err)))
-			return
-		}
-
-		foundUser = &newUserInfo
-
-		resp.WriteHeader(200)
-		resp.Write([]byte(fmt.Sprintf(`{"success": true, "username": "%s", "verified": %t, "apikey": "%s"}`, foundUser.Username, foundUser.Verified, foundUser.ApiKey)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(fmt.Sprintf(`{"success": true, "username": "%s", "verified": %t, "apikey": "%s"}`, userInfo.Username, userInfo.Verified, userInfo.ApiKey)))
-}
-
-func HandleSettings(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	userInfo, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in settings: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	newObject := SettingsReturn{
-		Success:  true,
-		Username: userInfo.Username,
-		Verified: userInfo.Verified,
-		Apikey:   userInfo.ApiKey,
-		Image:    userInfo.PublicProfile.GithubAvatar,
-	}
-
-	newjson, err := json.Marshal(newObject)
-	if err != nil {
-		log.Printf("[ERROR] Failed unmarshal in get settings: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed handling your user"}`)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write(newjson)
-}
-
-func SendHookResult(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in send hook results: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-	_ = user
-
-	location := strings.Split(request.URL.String(), "/")
-
-	var workflowId string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		workflowId = location[4]
-	}
-
-	if len(workflowId) != 32 {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "message": "ID not valid"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	hook, err := GetHook(ctx, workflowId)
-	if err != nil {
-		log.Printf("[WARNING] Failed getting hook %s (send): %s", workflowId, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("Body data error: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	log.Printf("SET the hook results for %s to %s", workflowId, body)
-	// FIXME - set the hook result in the DB somehow as interface{}
-	// FIXME - should the hook do the transform? Hmm
-
-	b, err := json.Marshal(hook)
-	if err != nil {
-		log.Printf("Failed marshalling: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(b))
-	return
-}
-
-func HandleGetHook(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in get hook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	location := strings.Split(request.URL.String(), "/")
-
-	var workflowId string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		workflowId = location[4]
-	}
-
-	if len(workflowId) != 36 {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "message": "ID not valid"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	hook, err := GetHook(ctx, workflowId)
-	if err != nil {
-		log.Printf("[WARNING] Failed getting hook %s (get hook): %s", workflowId, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Id != hook.Owner && user.Role != "scheduler" {
-		log.Printf("Wrong user (%s) for hook %s", user.Username, hook.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	b, err := json.Marshal(hook)
-	if err != nil {
-		log.Printf("Failed marshalling: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write([]byte(b))
-	return
-}
-
 
 func HandleKeyValueCheck(resp http.ResponseWriter, request *http.Request) {
 	cors := HandleCors(resp, request)
@@ -2099,945 +1015,11 @@ func HandleKeyValueCheck(resp http.ResponseWriter, request *http.Request) {
 	resp.Write(b)
 }
 
-// Used for swapping your own organization to a new one IF it's eligible
-func HandleChangeUserOrg(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	if project.Environment == "cloud" {
-		// Checking if it's a special region. All user-specific requests should
-		// go through shuffler.io and not subdomains
-		gceProject := os.Getenv("SHUFFLE_GCEPROJECT")
-		if gceProject != "shuffler" && len(gceProject) > 0 {
-			log.Printf("[DEBUG] Redirecting ORGCHANGE request to main site handler (shuffler.io)")
-			RedirectUserRequest(resp, request)
-			return
-		}
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed change org: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
-		return
-	}
-
-	type ReturnData struct {
-		OrgId     string `json:"org_id" datastore:"org_id"`
-		RegionUrl string `json:"region_url" datastore:"region_url"`
-	}
-
-	var tmpData ReturnData
-	err = json.Unmarshal(body, &tmpData)
-	if err != nil {
-		log.Printf("Failed unmarshalling test: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	ctx := GetContext(request)
-	foundOrg := false
-	for _, org := range user.Orgs {
-		if org == tmpData.OrgId {
-			foundOrg = true
-			break
-		}
-	}
-
-	if !foundOrg || tmpData.OrgId != fileId {
-		log.Printf("[WARNING] User swap to the org \"%s\" - access denied", tmpData.OrgId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "No permission to change to this org"}`))
-		return
-	}
-
-	org, err := GetOrg(ctx, tmpData.OrgId)
-	if err != nil {
-		log.Printf("[WARNING] Organization %s doesn't exist: %s", tmpData.OrgId, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Add instantswap of backend
-	// This could in theory be built out open source as well
-	regionUrl := ""
-	if project.Environment == "cloud" {
-		regionUrl = "https://shuffler.io"
-	}
-
-	if project.Environment == "cloud" && len(org.RegionUrl) > 0 && !strings.Contains(org.RegionUrl, "\"") {
-		regionUrl = org.RegionUrl
-	}
-
-	userFound := false
-	usr := User{}
-	for _, orgUsr := range org.Users {
-		//log.Printf("Usr: %#v", orgUsr)
-		if user.Id == orgUsr.Id {
-			usr = orgUsr
-			userFound = true
-			break
-		}
-	}
-
-	if !userFound {
-		log.Printf("[WARNING] User can't edit the org \"%s\" (2)", tmpData.OrgId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "No permission to edit this org"}`))
-		return
-	}
-
-	user.ActiveOrg = OrgMini{
-		Name: org.Name,
-		Id:   org.Id,
-		Role: usr.Role,
-	}
-
-	user.Role = usr.Role
-
-	err = SetUser(ctx, &user, false)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating user when changing org: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Cleanup cache for the user
-	DeleteCache(ctx, fmt.Sprintf("%s_workflows", user.Id))
-	DeleteCache(ctx, fmt.Sprintf("apps_%s", user.Id))
-	DeleteCache(ctx, fmt.Sprintf("user_%s", user.Username))
-	DeleteCache(ctx, fmt.Sprintf("user_%s", user.Id))
-
-	log.Printf("[INFO] User %s (%s) successfully changed org to %s (%s)", user.Username, user.Id, org.Name, org.Id)
-	resp.WriteHeader(200)
-	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully added new suborg. Refresh to see it.", "region_url": "%s"}`, regionUrl)))
-
-}
-
-func HandleCreateSubOrg(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in creating sub org: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Role != "admin" {
-		log.Printf("[WARNING] Not admin: %s (%s).", user.Username, user.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Not admin"}`))
-		return
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
-		return
-	}
-
-	type ReturnData struct {
-		OrgId string `json:"org_id" datastore:"org_id"`
-		Name  string `json:"name" datastore:"name"`
-	}
-
-	var tmpData ReturnData
-	err = json.Unmarshal(body, &tmpData)
-	if err != nil {
-		log.Printf("Failed unmarshalling test: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if len(tmpData.Name) < 3 {
-		log.Printf("[WARNING] Suborgname too short (min 3) %s", tmpData.Name)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Name must at least be 3 characters."}`))
-		return
-	}
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if tmpData.OrgId != user.ActiveOrg.Id || fileId != user.ActiveOrg.Id {
-		log.Printf("[WARNING] User can't edit the org \"%s\"", tmpData.OrgId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "No permission to edit this org"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	parentOrg, err := GetOrg(ctx, tmpData.OrgId)
-	if err != nil {
-		log.Printf("[WARNING] Organization %s doesn't exist: %s", tmpData.OrgId, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if len(parentOrg.ManagerOrgs) > 0 {
-		log.Printf("[WARNING] Organization %s can't have suborgs, as it's as suborg", tmpData.OrgId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Can't make suborg of suborg. Switch to a parent org to make another."}`))
-		return
-	}
-
-	if project.Environment == "cloud" {
-		if !parentOrg.SyncFeatures.MultiTenant.Active {
-			log.Printf("[WARNING] Org %s is not allowed to make a sub-organization: %s", tmpData.OrgId, err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false, "reason": "Sub-organizations require an active subscription with access to multi-tenancy. Contact support to try it out."}`))
-			return
-		}
-
-		/*
-			if parentOrg.SyncUsage.MultiTenant.Counter >= parentOrg.SyncFeatures.MultiTenant.Limit || len(parentOrg.ChildOrgs) > int(parentOrg.SyncFeatures.MultiTenant.Limit) {
-				log.Printf("[WARNING] Org %s is not allowed to make ANOTHER sub-organization. Limit reached!: %s", tmpData.OrgId, err)
-				resp.WriteHeader(401)
-				resp.Write([]byte(`{"success": false, "reason": "Your limit of sub-organizations has been reached. Contact support to increase."}`))
-				return
-			}
-		*/
-
-		parentOrg.SyncUsage.MultiTenant.Counter += 1
-		log.Printf("[DEBUG] Allowing suborg for %s because they have %d vs %d limit", parentOrg.Id, len(parentOrg.ChildOrgs), parentOrg.SyncFeatures.MultiTenant.Limit)
-	}
-
-	orgId := uuid.NewV4().String()
-	newOrg := Org{
-		Name:        tmpData.Name,
-		Description: fmt.Sprintf("Sub-org by user %s in parent-org %s", user.Username, parentOrg.Name),
-		Image:       parentOrg.Image,
-		Id:          orgId,
-		Org:         tmpData.Name,
-		Users: []User{
-			user,
-		},
-		Roles:     []string{"admin", "user"},
-		CloudSync: false,
-		ManagerOrgs: []OrgMini{
-			OrgMini{
-				Id:   tmpData.OrgId,
-				Name: parentOrg.Name,
-			},
-		},
-		CloudSyncActive: parentOrg.CloudSyncActive,
-		CreatorOrg:      tmpData.OrgId,
-	}
-	//SyncFeatures:    parentOrg.SyncFeatures,
-
-	parentOrg.ChildOrgs = append(parentOrg.ChildOrgs, OrgMini{
-		Name: tmpData.Name,
-		Id:   orgId,
-	})
-
-	err = SetOrg(ctx, newOrg, newOrg.Id)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting new org %s: %s", newOrg.Id, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	err = SetOrg(ctx, *parentOrg, parentOrg.Id)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating parent org %s: %s", newOrg.Id, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Update all admins to have access to this suborg
-	for _, loopUser := range parentOrg.Users {
-		if loopUser.Role != "admin" {
-			continue
-		}
-
-		if loopUser.Id == user.Id {
-			continue
-		}
-
-		foundUser, err := GetUser(ctx, loopUser.Id)
-		if err != nil {
-			log.Printf("[WARNING] User with Identifier %s doesn't exist: %s (update admins - create)", loopUser.Id, err)
-			continue
-		}
-
-		foundUser.Orgs = append(foundUser.Orgs, newOrg.Id)
-		err = SetUser(ctx, foundUser, false)
-		if err != nil {
-			log.Printf("[WARNING] Failed updating user when setting creating suborg (update admins - update): %s ", err)
-			continue
-		}
-	}
-
-	user.Orgs = append(user.Orgs, newOrg.Id)
-	log.Printf("[INFO] Usr: %#v (%d)", user.Orgs, len(user.Orgs))
-	err = SetUser(ctx, &user, false)
-	if err != nil {
-		log.Printf("[WARNING] Failed updating user when setting creating suborg: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	/*
-		if project.Environment != "cloud" {
-			log.Printf("[DEBUG] Starting cloud schedule for org %s (%s)", newOrg.Name, newOrg.Id)
-			interval := 15
-			log.Printf("[DEBUG] Should start schedule for org %s", newOrg.Name)
-			job := func() {
-				err := remoteOrgJobHandler(newOrg, interval)
-				if err != nil {
-					log.Printf("[ERROR] Failed request with remote org setup (3): %s", err)
-				}
-			}
-
-			jobret, err := newscheduler.Every(int(interval)).Seconds().NotImmediately().Run(job)
-			if err != nil {
-				log.Printf("[CRITICAL] Failed to schedule new org: %s", err)
-			}
-		}
-	*/
-
-	log.Printf("[INFO] User %s SUCCESSFULLY ADDED child org %s (%s) for parent %s (%s)", user.Username, newOrg.Name, newOrg.Id, parentOrg.Name, parentOrg.Id)
-	resp.WriteHeader(200)
-	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully updated org"}`)))
-
-}
-
-func HandleEditOrg(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in edit org: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Role != "admin" {
-		log.Printf("[WARNING] Not admin: %s (%s).", user.Username, user.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Not admin"}`))
-		return
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
-		return
-	}
-
-	type ReturnData struct {
-		Tutorial    string    `json:"tutorial" datastore:"tutorial"`
-		Name        string    `json:"name" datastore:"name"`
-		Image       string    `json:"image" datastore:"image"`
-		CompanyType string    `json:"company_type" datastore:"company_type"`
-		Description string    `json:"description" datastore:"description"`
-		OrgId       string    `json:"org_id" datastore:"org_id"`
-		Priority    string    `json:"priority" datastore:"priority"`
-		Defaults    Defaults  `json:"defaults" datastore:"defaults"`
-		SSOConfig   SSOConfig `json:"sso_config" datastore:"sso_config"`
-	}
-
-	var tmpData ReturnData
-	err = json.Unmarshal(body, &tmpData)
-	if err != nil {
-		log.Printf("Failed unmarshalling test: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-	//log.Printf("SSO: %#v", tmpData.SSOConfig)
-
-	var fileId string
-	location := strings.Split(request.URL.String(), "/")
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			log.Printf("Path too short: %d", len(location))
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if tmpData.OrgId != user.ActiveOrg.Id || fileId != user.ActiveOrg.Id {
-		log.Printf("[WARNING] User can't edit org %#v (active: %#v)", fileId, user.ActiveOrg.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "No permission to edit this org"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	org, err := GetOrg(ctx, tmpData.OrgId)
-	if err != nil {
-		log.Printf("[WARNING] Organization doesn't exist: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	admin := false
-	userFound := false
-	for _, inneruser := range org.Users {
-		if inneruser.Id == user.Id {
-			userFound = true
-			if inneruser.Role == "admin" {
-				admin = true
-			}
-
-			break
-		}
-	}
-
-	if !userFound {
-		log.Printf("User %s doesn't exist in organization for edit %s", user.Id, org.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if !admin {
-		log.Printf("User %s doesn't have edit rights to %s", user.Id, org.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	sendOrgUpdaterHook := false
-	if len(tmpData.Image) > 0 {
-		org.Image = tmpData.Image
-	}
-
-	if len(tmpData.Name) > 0 {
-		org.Name = tmpData.Name
-	}
-
-	if len(tmpData.Description) > 0 {
-		org.Description = tmpData.Description
-	}
-
-	if len(tmpData.Defaults.AppDownloadRepo) > 0 || len(tmpData.Defaults.AppDownloadBranch) > 0 || len(tmpData.Defaults.WorkflowDownloadRepo) > 0 || len(tmpData.Defaults.WorkflowDownloadBranch) > 0 || len(tmpData.Defaults.NotificationWorkflow) > 0 {
-		org.Defaults = tmpData.Defaults
-	}
-
-	if len(tmpData.CompanyType) > 0 {
-		org.CompanyType = tmpData.CompanyType
-
-		if len(org.CompanyType) == 0 {
-			sendOrgUpdaterHook = true
-		}
-	}
-
-	if len(tmpData.Tutorial) > 0 {
-		if tmpData.Tutorial == "welcome" {
-			sendOrgUpdaterHook = true
-		}
-	}
-
-	if len(tmpData.Priority) > 0 {
-		if len(org.MainPriority) == 0 {
-			org.MainPriority = tmpData.Priority
-			sendOrgUpdaterHook = true
-		}
-
-		found := false
-		for _, prio := range org.Priorities {
-			if prio.Name == tmpData.Priority {
-				found = true
-			}
-		}
-
-		if !found {
-			org.Priorities = append(org.Priorities, Priority{
-				Name:        tmpData.Priority,
-				Description: fmt.Sprintf("Priority %s decided by user.", tmpData.Priority),
-				Type:        "usecases",
-				Active:      true,
-				URL:         fmt.Sprintf("/usecases"),
-			})
-		}
-	}
-
-	//if len(tmpData.SSOConfig) > 0 {
-	if len(tmpData.SSOConfig.SSOEntrypoint) > 0 || len(tmpData.SSOConfig.OpenIdClientId) > 0 || len(tmpData.SSOConfig.OpenIdClientSecret) > 0 || len(tmpData.SSOConfig.OpenIdAuthorization) > 0 || len(tmpData.SSOConfig.OpenIdToken) > 0 {
-		org.SSOConfig = tmpData.SSOConfig
-	}
-
-	if len(tmpData.SSOConfig.SSOCertificate) > 0 {
-		savedCert := fixCertificate(tmpData.SSOConfig.SSOCertificate)
-
-		log.Printf("[INFO] Stripped down cert from %d to %d", len(tmpData.SSOConfig.SSOCertificate), len(savedCert))
-
-		org.SSOConfig.SSOCertificate = savedCert
-	}
-
-	if len(org.Defaults.NotificationWorkflow) > 0 && len(org.Defaults.NotificationWorkflow) != 36 {
-		log.Printf("[WARNING] Notification Workflow ID %s is not valid.", org.Defaults.NotificationWorkflow)
-	}
-
-	// Built a system around this now, which checks for the actual org. Only works onprem so far.
-	// if requestdata.Environment == "cloud" && project.Environment != "cloud" {
-	//if project.Environment != "cloud" && len(org.SSOConfig.SSOEntrypoint) > 0 && len(org.ManagerOrgs) == 0 {
-	//	//log.Printf("[INFO] Should set SSO entrypoint to %s", org.SSOConfig.SSOEntrypoint)
-	//	SSOUrl = org.SSOConfig.SSOEntrypoint
-	//}
-
-	//log.Printf("Org: %#v", org)
-	err = SetOrg(ctx, *org, org.Id)
-	if err != nil {
-		log.Printf("User %s doesn't have edit rights to %s", user.Id, org.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Only send for cloud
-	if sendOrgUpdaterHook && project.Environment == "cloud" {
-		signupWebhook := os.Getenv("WEBSITE_ORG_WEBHOOK")
-		if strings.HasPrefix(signupWebhook, "http") {
-			curIndex := -1
-			for orgIndex, orguser := range org.Users {
-				if orguser.Id == user.Id {
-					curIndex = orgIndex
-				}
-			}
-
-			if curIndex >= 0 {
-				user.Password = ""
-				user.Session = ""
-				user.ApiKey = ""
-				user.LoginInfo = []LoginInfo{}
-				user.PrivateApps = []WorkflowApp{}
-
-				org.Users[curIndex] = user
-			}
-
-			mappedData, err := json.Marshal(org)
-			if err != nil {
-				log.Printf("[WARNING] Marshal error for org sending: %s", err)
-			} else {
-				req, err := http.NewRequest(
-					"POST",
-					signupWebhook,
-					bytes.NewBuffer(mappedData),
-				)
-
-				client := &http.Client{
-					Timeout: 3 * time.Second,
-				}
-
-				req.Header.Add("Content-Type", "application/json")
-				res, err := client.Do(req)
-				if err != nil {
-					log.Printf("[ERROR] Failed request to signup webhook FOR ORG (2): %s", err)
-				} else {
-					log.Printf("[INFO] Successfully ran org priority webhook")
-				}
-
-				_ = res
-			}
-		}
-	}
-
-	log.Printf("[INFO] Successfully updated org %s (%s)", org.Name, org.Id)
-	resp.WriteHeader(200)
-	resp.Write([]byte(fmt.Sprintf(`{"success": true, "reason": "Successfully updated org"}`)))
-
-}
-
-// Starts a new webhook
-func HandleNewHook(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in set new hook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Role == "org-reader" {
-		log.Printf("[WARNING] Org-reader doesn't have access to make new hook: %s (%s)", user.Username, user.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Read only user"}`))
-		return
-	}
-
-	type requestData struct {
-		Type           string `json:"type"`
-		Description    string `json:"description"`
-		Id             string `json:"id"`
-		Name           string `json:"name"`
-		Workflow       string `json:"workflow"`
-		Start          string `json:"start"`
-		Environment    string `json:"environment"`
-		Auth           string `json:"auth"`
-		CustomResponse string `json:"custom_response"`
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("[WARNING] Body data error in webhook set: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	var requestdata requestData
-	err = json.Unmarshal([]byte(body), &requestdata)
-	if err != nil {
-		log.Printf("[WARNING] Failed unmarshaling inputdata for webhook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	newId := requestdata.Id
-	if len(newId) != 36 {
-		log.Printf("[WARNING] Bad webhook ID: %s", newId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Invalid Webhook ID: bad formatting"}`))
-		return
-	}
-
-	if requestdata.Id == "" || requestdata.Name == "" {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Required fields id and name can't be empty"}`))
-		return
-
-	}
-
-	validTypes := []string{
-		"webhook",
-	}
-
-	isTypeValid := false
-	for _, thistype := range validTypes {
-		if requestdata.Type == thistype {
-			isTypeValid = true
-			break
-		}
-	}
-
-	if !(isTypeValid) {
-		log.Printf("Type %s is not valid. Try any of these: %s", requestdata.Type, strings.Join(validTypes, ", "))
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Let remote endpoint handle access checks (shuffler.io)
-	baseUrl := "https://shuffler.io"
-	if len(os.Getenv("SHUFFLE_GCEPROJECT")) > 0 && len(os.Getenv("SHUFFLE_GCEPROJECT_LOCATION")) > 0 {
-		baseUrl = fmt.Sprintf("https://%s.%s.r.appspot.com", os.Getenv("SHUFFLE_GCEPROJECT"), os.Getenv("SHUFFLE_GCEPROJECT_LOCATION"))
-	}
-
-	if len(os.Getenv("SHUFFLE_CLOUDRUN_URL")) > 0 {
-		baseUrl = os.Getenv("SHUFFLE_CLOUDRUN_URL")
-	}
-
-	currentUrl := fmt.Sprintf("%s/api/v1/hooks/webhook_%s", baseUrl, newId)
-	startNode := requestdata.Start
-	if requestdata.Environment == "cloud" && project.Environment != "cloud" {
-		// https://shuffler.io/v1/hooks/webhook_80184973-3e82-4852-842e-0290f7f34d7c
-		log.Printf("[INFO] Should START a cloud webhook for url %s for startnode %s", currentUrl, startNode)
-		org, err := GetOrg(ctx, user.ActiveOrg.Id)
-		if err != nil {
-			log.Printf("Failed finding org %s: %s", org.Id, err)
-			return
-		}
-
-		action := CloudSyncJob{
-			Type:          "webhook",
-			Action:        "start",
-			OrgId:         org.Id,
-			PrimaryItemId: newId,
-			SecondaryItem: startNode,
-			ThirdItem:     requestdata.Workflow,
-			FourthItem:    requestdata.Auth,
-		}
-
-		err = executeCloudAction(action, org.SyncConfig.Apikey)
-		if err != nil {
-			log.Printf("[WARNING] Failed cloud action START webhook execution: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "%s"}`, err)))
-			return
-		} else {
-			log.Printf("[INFO] Successfully set up cloud action schedule")
-		}
-	}
-
-	hook := Hook{
-		Id:        newId,
-		Start:     startNode,
-		Workflows: []string{requestdata.Workflow},
-		Info: Info{
-			Name:        requestdata.Name,
-			Description: requestdata.Description,
-			Url:         fmt.Sprintf("%s/api/v1/hooks/webhook_%s", baseUrl, newId),
-		},
-		Type:   "webhook",
-		Owner:  user.Username,
-		Status: "uninitialized",
-		Actions: []HookAction{
-			HookAction{
-				Type:  "workflow",
-				Name:  requestdata.Name,
-				Id:    requestdata.Workflow,
-				Field: "",
-			},
-		},
-		Running:        false,
-		OrgId:          user.ActiveOrg.Id,
-		Environment:    requestdata.Environment,
-		Auth:           requestdata.Auth,
-		CustomResponse: requestdata.CustomResponse,
-	}
-
-	hook.Status = "running"
-	hook.Running = true
-	err = SetHook(ctx, hook)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting hook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	log.Printf("[INFO] Set up a new hook with ID %s and environment %s", newId, hook.Environment)
-	resp.WriteHeader(200)
-	resp.Write([]byte(`{"success": true}`))
-}
-
-func HandleDeleteHook(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[WARNING] Api authentication failed in delete hook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Role == "org-reader" {
-		log.Printf("[WARNING] Org-reader doesn't have access to delete hook: %s (%s)", user.Username, user.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Read only user"}`))
-		return
-	}
-
-	location := strings.Split(request.URL.String(), "/")
-
-	var fileId string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if len(fileId) != 36 {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Workflow ID when deleting hook is not valid"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	hook, err := GetHook(ctx, fileId)
-	if err != nil {
-		log.Printf("[WARNING] Failed getting hook %s (delete): %s", fileId, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if user.Id != hook.Owner && user.ActiveOrg.Id != hook.OrgId {
-		log.Printf("[WARNING] Wrong user (%s) for workflow %s", user.Username, hook.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if len(hook.Workflows) > 0 {
-		//err = increaseStatisticsField(ctx, "total_workflow_triggers", hook.Workflows[0], -1, user.ActiveOrg.Id)
-		//if err != nil {
-		//	log.Printf("Failed to increase total workflows: %s", err)
-		//}
-	}
-
-	hook.Status = "stopped"
-	err = SetHook(ctx, *hook)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting hook: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if hook.Environment == "cloud" && project.Environment != "cloud" {
-		log.Printf("[INFO] Should STOP cloud webhook https://shuffler.io/api/v1/hooks/webhook_%s", hook.Id)
-		org, err := GetOrg(ctx, user.ActiveOrg.Id)
-		if err != nil {
-			log.Printf("Failed finding org %s: %s", org.Id, err)
-			return
-		}
-
-		action := CloudSyncJob{
-			Type:          "webhook",
-			Action:        "stop",
-			OrgId:         org.Id,
-			PrimaryItemId: hook.Id,
-		}
-
-		if len(hook.Workflows) > 0 {
-			action.SecondaryItem = hook.Workflows[0]
-		}
-
-		err = executeCloudAction(action, org.SyncConfig.Apikey)
-		if err != nil {
-			log.Printf("Failed cloud action STOP execution: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "%s"}`, err)))
-			return
-		}
-		// https://shuffler.io/v1/hooks/webhook_80184973-3e82-4852-842e-0290f7f34d7c
-	}
-
-	err = DeleteKey(ctx, "hooks", fileId)
-	if err != nil {
-		log.Printf("[WARNING] Error deleting hook %s for %s: %s", fileId, user.Username, err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed deleting the hook."}`))
-		return
-	}
-
-	log.Printf("[INFO] Successfully deleted webhook %s", fileId)
-	resp.WriteHeader(200)
-	resp.Write([]byte(`{"success": true, "reason": "Stopped webhook"}`))
-}
-
 func ParseVersions(versions []string) []string {
 	log.Printf("Versions: %#v", versions)
 
 	//versions = sort.Sort(semver.Collection(versions))
 	return versions
-}
-
-// Finds the child nodes of a node in execution and returns them
-// Used if e.g. a node in a branch is exited, and all children have to be stopped
-func FindChildNodes(workflowExecution WorkflowExecution, nodeId string) []string {
-	//log.Printf("\nNODE TO FIX: %s\n\n", nodeId)
-	allChildren := []string{nodeId}
-	//log.Printf("\n\n")
-
-	// 1. Find children of this specific node
-	// 2. Find the children of those nodes etc.
-	// 3. Sort it in the right order to handle merges properly
-	for _, branch := range workflowExecution.Workflow.Branches {
-		if branch.SourceID == nodeId {
-			//log.Printf("NODE: %s, SRC: %s, CHILD: %s\n", nodeId, branch.SourceID, branch.DestinationID)
-			allChildren = append(allChildren, branch.DestinationID)
-
-			childNodes := FindChildNodes(workflowExecution, branch.DestinationID)
-			for _, bottomChild := range childNodes {
-				found := false
-
-				for _, topChild := range allChildren {
-					if topChild == bottomChild {
-						found = true
-						break
-					}
-				}
-
-				if !found {
-					allChildren = append(allChildren, bottomChild)
-				}
-			}
-		}
-	}
-
-	// Remove potential duplicates
-	newNodes := []string{}
-	for _, tmpnode := range allChildren {
-		found := false
-		for _, newnode := range newNodes {
-			if newnode == tmpnode {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			newNodes = append(newNodes, tmpnode)
-		}
-	}
-
-	return newNodes
 }
 
 // Can't just regex out stuff due to unicode problems with other languages
@@ -3381,188 +1363,6 @@ func ValidateSwagger(resp http.ResponseWriter, request *http.Request) {
 	// save the openapi ID
 	resp.WriteHeader(422)
 	resp.Write([]byte(`{"success": false}`))
-}
-
-// Recursively finds child nodes inside sub workflows
-func GetReplacementNodes(ctx context.Context, execution WorkflowExecution, trigger Trigger, lastTriggerName string) ([]Action, []Branch, string) {
-	if execution.ExecutionOrg == "" {
-		execution.ExecutionOrg = execution.Workflow.OrgId
-	}
-
-	selectedWorkflow := ""
-	workflowAction := ""
-	for _, param := range trigger.Parameters {
-		if param.Name == "workflow" {
-			selectedWorkflow = param.Value
-		}
-
-		if param.Name == "startnode" {
-			workflowAction = param.Value
-		}
-	}
-
-	if len(selectedWorkflow) == 0 {
-		return []Action{}, []Branch{}, ""
-	}
-
-	// Authenticating and such
-	workflow, err := GetWorkflow(ctx, selectedWorkflow)
-	if err != nil {
-		return []Action{}, []Branch{}, ""
-	}
-
-	orgFound := false
-	if workflow.ExecutingOrg.Id == execution.ExecutionOrg {
-		orgFound = true
-	} else if workflow.OrgId == execution.ExecutionOrg {
-		orgFound = true
-	} else {
-		for _, org := range workflow.Org {
-			if org.Id == execution.ExecutionOrg {
-				orgFound = true
-				break
-			}
-		}
-	}
-
-	if !orgFound {
-		log.Printf("[WARNING] Auth for subflow is bad. %s (orig) vs %s", execution.ExecutionOrg, workflow.OrgId)
-		return []Action{}, []Branch{}, ""
-	}
-
-	//childNodes = FindChildNodes(workflowExecution, actionResult.Action.ID)
-	//log.Printf("FIND CHILDNODES OF STARTNODE %s", workflowAction)
-	workflowExecution := WorkflowExecution{
-		Workflow: *workflow,
-	}
-
-	childNodes := FindChildNodes(workflowExecution, workflowAction)
-	//log.Printf("Found %d childnodes of %s", len(childNodes), workflowAction)
-	newActions := []Action{}
-	branches := []Branch{}
-
-	// FIXME: Bad lastnode check. Need to go to the bottom of workflows and check max steps away from parent
-	lastNode := ""
-	for _, nodeId := range childNodes {
-		for _, action := range workflow.Actions {
-			if nodeId == action.ID {
-				newActions = append(newActions, action)
-				break
-			}
-		}
-
-		for _, branch := range workflow.Branches {
-			if branch.SourceID == nodeId {
-				branches = append(branches, branch)
-			}
-		}
-
-		lastNode = nodeId
-	}
-
-	found := false
-	for actionIndex, action := range newActions {
-		if lastNode == action.ID {
-			//actions[actionIndex].Name = trigger.Name
-			newActions[actionIndex].Label = lastTriggerName
-			//trigger.Label
-			found = true
-		}
-	}
-
-	if !found {
-		log.Printf("SHOULD CHECK TRIGGERS FOR LASTNODE!")
-	}
-
-	log.Printf("[INFO] Found %d actions and %d branches in subflow", len(newActions), len(branches))
-	if len(newActions) == len(childNodes) {
-		return newActions, branches, lastNode
-	} else {
-		log.Printf("\n\n[WARNING] Bad length of actions and nodes in subflow (subsubflow?): %d vs %d", len(newActions), len(childNodes))
-
-		// Adding information about triggers if subflow
-		changed := false
-		for _, nodeId := range childNodes {
-			for triggerIndex, trigger := range workflow.Triggers {
-				if trigger.AppName == "Shuffle Workflow" {
-					if nodeId == trigger.ID {
-						replaceActions := false
-						workflowAction := ""
-						for _, param := range trigger.Parameters {
-							if param.Name == "argument" && !strings.Contains(param.Value, ".#") {
-								replaceActions = true
-							}
-
-							if param.Name == "startnode" {
-								workflowAction = param.Value
-							}
-						}
-
-						if replaceActions {
-							replacementNodes, newBranches, lastNode := GetReplacementNodes(ctx, workflowExecution, trigger, lastTriggerName)
-							log.Printf("SUB REPLACEMENTS: %d, %d", len(replacementNodes), len(newBranches))
-							log.Printf("\n\nNEW LASTNODE: %s\n\n", lastNode)
-							if len(replacementNodes) > 0 {
-								//workflowExecution.Workflow.Actions = append(workflowExecution.Workflow.Actions, action)
-
-								//lastnode = replacementNodes[0]
-								// Have to validate in case it's the same workflow and such
-								for _, action := range replacementNodes {
-									found := false
-									for subActionIndex, subaction := range newActions {
-										if subaction.ID == action.ID {
-											found = true
-											//newActions[subActionIndex].Name = action.Name
-											newActions[subActionIndex].Label = action.Label
-											break
-										}
-									}
-
-									if !found {
-										action.SubAction = true
-										newActions = append(newActions, action)
-									}
-								}
-
-								for _, branch := range newBranches {
-									workflowExecution.Workflow.Branches = append(workflowExecution.Workflow.Branches, branch)
-								}
-
-								// Append branches:
-								// parent -> new inner node (FIRST one)
-								for branchIndex, branch := range workflowExecution.Workflow.Branches {
-									if branch.DestinationID == trigger.ID {
-										log.Printf("REPLACE DESTINATION WITH %s!!", workflowAction)
-										workflowExecution.Workflow.Branches[branchIndex].DestinationID = workflowAction
-										branches = append(branches, workflowExecution.Workflow.Branches[branchIndex])
-									}
-
-									if branch.SourceID == trigger.ID {
-										log.Printf("REPLACE SOURCE WITH LASTNODE %s!!", lastNode)
-										workflowExecution.Workflow.Branches[branchIndex].SourceID = lastNode
-										branches = append(branches, workflowExecution.Workflow.Branches[branchIndex])
-									}
-								}
-
-								// Remove the trigger
-								workflowExecution.Workflow.Triggers = append(workflowExecution.Workflow.Triggers[:triggerIndex], workflowExecution.Workflow.Triggers[triggerIndex+1:]...)
-								workflow.Triggers = append(workflow.Triggers[:triggerIndex], workflow.Triggers[triggerIndex+1:]...)
-								changed = true
-							}
-						}
-					}
-				}
-
-			}
-		}
-
-		log.Printf("NEW ACTION LENGTH %d, Branches: %d. LASTNODE: %s\n\n", len(newActions), len(branches), lastNode)
-		if changed {
-			return newActions, branches, lastNode
-		}
-	}
-
-	return []Action{}, []Branch{}, ""
 }
 
 /*
@@ -3950,225 +1750,6 @@ func HandleSetCacheKey(resp http.ResponseWriter, request *http.Request) {
 	resp.Write(b)
 }
 
-// Checks authentication string for Webhooks
-func CheckHookAuth(request *http.Request, auth string) error {
-	if len(auth) == 0 {
-		return nil
-	}
-
-	authSplit := strings.Split(auth, "\n")
-	for _, line := range authSplit {
-		lineSplit := strings.Split(line, "=")
-		if strings.Contains(line, ":") {
-			lineSplit = strings.Split(line, "=")
-		}
-
-		if len(lineSplit) == 2 {
-			validationHeader := strings.ToLower(strings.TrimSpace(lineSplit[0]))
-			found := false
-			for key, value := range request.Header {
-				if strings.ToLower(key) == validationHeader && len(value) > 0 {
-					//log.Printf("FOUND KEY %#v. Value: %s", validationHeader, value)
-					if value[0] == strings.TrimSpace(lineSplit[1]) {
-						found = true
-						break
-					}
-				}
-			}
-
-			if !found {
-				return errors.New(fmt.Sprintf("Missing or bad header: %#v", validationHeader))
-			}
-
-			//log.Printf("Find header %#v", validationHeader)
-			//itemHeader := request.Header[validationHeader]
-			//log.Printf("LINE: %s. Header: %s", line, itemHeader)
-		} else {
-			log.Printf("[WARNING] Bad auth line: %s. NOT checking auth.", line)
-		}
-	}
-
-	//return errors.New("Bad auth!")
-	return nil
-}
-
-// Fileid = the app to execute
-// Body = The action body received from the user to test.
-func PrepareSingleAction(ctx context.Context, user User, fileId string, body []byte) (WorkflowExecution, error) {
-	var action Action
-	workflowExecution := WorkflowExecution{}
-	err := json.Unmarshal(body, &action)
-	if err != nil {
-		log.Printf("[WARNING] Failed action single execution unmarshaling: %s", err)
-		return workflowExecution, err
-	}
-
-	/*
-		if len(workflow.Name) > 0 || len(workflow.Owner) > 0 || len(workflow.OrgId) > 0 || len(workflow.Actions) != 1 {
-			log.Printf("[WARNING] Bad length for some characteristics in single execution of %s", fileId)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-	*/
-
-	if fileId != action.AppID {
-		log.Printf("[WARNING] Bad appid in single execution of App %s", fileId)
-		return workflowExecution, err
-	}
-
-	if len(action.ID) == 0 {
-		action.ID = uuid.NewV4().String()
-	}
-
-	app, err := GetApp(ctx, fileId, user, false)
-	if err != nil {
-		log.Printf("[WARNING] Error getting app (execute SINGLE workflow): %s", fileId)
-		return workflowExecution, err
-	}
-
-	if app.Authentication.Required && len(action.AuthenticationId) == 0 {
-		log.Printf("[WARNING] Tried to execute SINGLE %s WITHOUT auth (missing)", app.Name)
-
-		found := false
-		for _, param := range action.Parameters {
-			if param.Configuration {
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			return workflowExecution, errors.New("You must authenticate the app first")
-		}
-	}
-
-	newParams := []WorkflowAppActionParameter{}
-	if len(action.AuthenticationId) > 0 {
-		log.Printf("[INFO] Adding auth in single execution!")
-		curAuth, err := GetWorkflowAppAuthDatastore(ctx, action.AuthenticationId)
-
-		if err != nil {
-			log.Printf("[WARNING] Failed getting authentication for your org: %s", err)
-			return workflowExecution, err
-		}
-
-		if user.ActiveOrg.Id != curAuth.OrgId {
-			log.Printf("[WARNING] User %s tried to use bad auth %s", user.Id, action.AuthenticationId)
-			return workflowExecution, err
-		}
-
-		if curAuth.Encrypted {
-			for _, field := range curAuth.Fields {
-				parsedKey := fmt.Sprintf("%s_%d_%s_%s", curAuth.OrgId, curAuth.Created, curAuth.Label, field.Key)
-				newValue, err := HandleKeyDecryption([]byte(field.Value), parsedKey)
-				if err != nil {
-					log.Printf("[ERROR] Failed decryption for %s: %s", field.Key, err)
-					break
-				}
-
-				if field.Key == "url" {
-					//log.Printf("Value2 (%s): %s", field.Key, string(newValue))
-					if strings.HasSuffix(string(newValue), "/") {
-						newValue = []byte(string(newValue)[0 : len(newValue)-1])
-					}
-
-					//log.Printf("Value2 (%s): %s", field.Key, string(newValue))
-				}
-
-				newParam := WorkflowAppActionParameter{
-					Name:  field.Key,
-					ID:    action.AuthenticationId,
-					Value: string(newValue),
-				}
-
-				newParams = append(newParams, newParam)
-			}
-		} else {
-			//log.Printf("[INFO] AUTH IS NOT ENCRYPTED - attempting auto-encrypting if key is set!")
-			//err = SetWorkflowAppAuthDatastore(ctx, curAuth, curAuth.Id)
-			//if err != nil {
-			//	log.Printf("[WARNING] Failed running encryption during execution: %s", err)
-			//}
-			for _, auth := range curAuth.Fields {
-
-				newParam := WorkflowAppActionParameter{
-					Name:  auth.Key,
-					ID:    action.AuthenticationId,
-					Value: auth.Value,
-				}
-
-				newParams = append(newParams, newParam)
-			}
-		}
-
-		// Rebuild params with the right data. This is to prevent issues on the frontend
-
-		action.Parameters = newParams
-	}
-
-	for _, param := range action.Parameters {
-		if param.Required && len(param.Value) == 0 {
-			//log.Printf("Param: %#v", param)
-
-			if param.Name == "username_basic" {
-				param.Name = "username"
-			} else if param.Name == "password_basic" {
-				param.Name = "password"
-			}
-
-			param.Name = strings.Replace(param.Name, "_", " ", -1)
-			param.Name = strings.Title(param.Name)
-
-			value := fmt.Sprintf("Param %s can't be empty. Please fill all required parameters (orange outline). If you don't know the value, input space in the field.", param.Name)
-			log.Printf("[WARNING] During single exec: %s", value)
-			return workflowExecution, errors.New(value)
-		}
-
-		newParams = append(newParams, param)
-	}
-
-	action.Sharing = app.Sharing
-	action.Public = app.Public
-	action.Generated = app.Generated
-	action.Parameters = newParams
-
-	workflow := Workflow{
-		Actions: []Action{
-			action,
-		},
-		Start: action.ID,
-		ID:    uuid.NewV4().String(),
-	}
-
-	//log.Printf("Sharing: %#v, Public: %#v, Generated: %#v. Start: %#v", action.Sharing, action.Public, action.Generated, workflow.Start)
-
-	workflowExecution = WorkflowExecution{
-		Workflow:      workflow,
-		Start:         workflow.Start,
-		ExecutionId:   uuid.NewV4().String(),
-		WorkflowId:    workflow.ID,
-		StartedAt:     int64(time.Now().Unix()),
-		CompletedAt:   0,
-		Authorization: uuid.NewV4().String(),
-		Status:        "EXECUTING",
-	}
-
-	if user.ActiveOrg.Id != "" {
-		workflow.ExecutingOrg = user.ActiveOrg
-		workflowExecution.ExecutionOrg = user.ActiveOrg.Id
-		workflowExecution.OrgId = user.ActiveOrg.Id
-	}
-
-	err = SetWorkflowExecution(ctx, workflowExecution, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed handling single execution setup: %s", err)
-		return workflowExecution, err
-	}
-
-	return workflowExecution, nil
-}
-
 func HandleRetValidation(ctx context.Context, workflowExecution WorkflowExecution) []byte {
 	cnt := 0
 	type retStruct struct {
@@ -4223,8 +1804,6 @@ func HandleRetValidation(ctx context.Context, workflowExecution WorkflowExecutio
 	return returnBytes
 }
 
-
-
 func md5sum(data []byte) string {
 	hasher := md5.New()
 	hasher.Write(data)
@@ -4252,884 +1831,6 @@ func fixCertificate(parsedX509Key string) string {
 	return parsedX509Key
 }
 
-// Example implementation of SSO, including a redirect for the user etc
-// Should make this stuff only possible after login
-func HandleOpenId(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	//https://dev-18062475.okta.com/oauth2/default/v1/authorize?client_id=oa3romteykJ2aMgx5d7&response_type=code&scope=openid&redirect_uri=http%3A%2F%2Flocalhost%3A5002%2Fapi%2Fv1%2Flogin_openid&state=state-296bc9a0-a2a2-4a57-be1a-d0e2fd9bb601&code_challenge_method=S256&code_challenge=codechallenge
-	// http://localhost:5002/api/v1/login_openid?code=rrm8BS8eUIYpQWnoM_Lzh_QoT3-EwQ2c9YkjRcJWqk4&state=state-296bc9a0-a2a2-4a57-be1a-d0e2fd9bb601
-	// http://localhost:5001/api/v1/login_openid#id_token=asdasd&session_state=asde9d78d8-6535-45fe-848d-0efa9f119595
-
-	//code -> Token
-	ctx := GetContext(request)
-
-	skipValidation := false
-	openidUser := OpenidUserinfo{}
-	org := &Org{}
-	code := request.URL.Query().Get("code")
-	if len(code) == 0 {
-		// Check id_token grant info
-		if request.Method == "POST" {
-			body, err := ioutil.ReadAll(request.Body)
-			if err != nil {
-				resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No code or id_token specified - body read error in POST"}`)))
-				resp.WriteHeader(401)
-				return
-			}
-
-			stateSplit := strings.Split(string(body), "&")
-			for _, innerstate := range stateSplit {
-				itemsplit := strings.Split(innerstate, "=")
-
-				if len(itemsplit) <= 1 {
-					log.Printf("[WARNING] No key:value: %s", innerstate)
-					continue
-				}
-
-				if itemsplit[0] == "id_token" {
-					token, err := VerifyIdToken(ctx, itemsplit[1])
-					if err != nil {
-						log.Printf("[ERROR] Bad ID token provided: %s", err)
-						resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Bad ID token provided"}`)))
-						resp.WriteHeader(401)
-						return
-					}
-
-					log.Printf("[DEBUG] Validated - token: %s!", token)
-					openidUser.Sub = token.Sub
-					org = &token.Org
-					skipValidation = true
-
-					break
-				}
-			}
-		}
-
-		if !skipValidation {
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No code specified"}`)))
-			resp.WriteHeader(401)
-			return
-		}
-	}
-
-	if !skipValidation {
-		state := request.URL.Query().Get("state")
-		if len(state) == 0 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No state specified"}`)))
-			return
-		}
-
-		stateBase, err := base64.StdEncoding.DecodeString(state)
-		if err != nil {
-			log.Printf("[ERROR] Failed base64 decode OpenID state: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed base64 decoding of state"}`)))
-			return
-		}
-
-		log.Printf("State: %s", stateBase)
-		foundOrg := ""
-		foundRedir := ""
-		foundChallenge := ""
-		stateSplit := strings.Split(string(stateBase), "&")
-		for _, innerstate := range stateSplit {
-			itemsplit := strings.Split(innerstate, "=")
-			//log.Printf("Itemsplit: %#v", itemsplit)
-			if len(itemsplit) <= 1 {
-				log.Printf("[WARNING] No key:value: %s", innerstate)
-				continue
-			}
-
-			if itemsplit[0] == "org" {
-				foundOrg = strings.TrimSpace(itemsplit[1])
-			}
-
-			if itemsplit[0] == "redirect" {
-				foundRedir = strings.TrimSpace(itemsplit[1])
-			}
-
-			if itemsplit[0] == "challenge" {
-				foundChallenge = strings.TrimSpace(itemsplit[1])
-			}
-		}
-
-		//log.Printf("Challenge len2: %d", len(foundChallenge))
-
-		if len(foundOrg) == 0 {
-			log.Printf("[ERROR] No org specified in state")
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No org specified in state"}`)))
-			return
-		}
-
-		org, err = GetOrg(ctx, foundOrg)
-		if err != nil {
-			log.Printf("[WARNING] Error getting org in OpenID: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false, "reason": "Couldn't find the org for sign-in in Shuffle"}`))
-			return
-		}
-
-		clientId := org.SSOConfig.OpenIdClientId
-		tokenUrl := org.SSOConfig.OpenIdToken
-		if len(tokenUrl) == 0 {
-			log.Printf("[ERROR] No token URL specified for OpenID")
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No token URL specified. Please make sure to specify a token URL in the /admin panel in Shuffle for OpenID Connect"}`)))
-			return
-		}
-
-		//log.Printf("Challenge: %s", foundChallenge)
-		body, err := RunOpenidLogin(ctx, clientId, tokenUrl, foundRedir, code, foundChallenge, org.SSOConfig.OpenIdClientSecret)
-		if err != nil {
-			log.Printf("[WARNING] Error with body read of OpenID Connect: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		openid := OpenidResp{}
-		err = json.Unmarshal(body, &openid)
-		if err != nil {
-			log.Printf("[WARNING] Error in Openid marshal: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		// Automated replacement
-		userInfoUrlSplit := strings.Split(org.SSOConfig.OpenIdAuthorization, "/")
-		userinfoEndpoint := strings.Join(userInfoUrlSplit[0:len(userInfoUrlSplit)-1], "/") + "/userinfo"
-		//userinfoEndpoint := strings.Replace(org.SSOConfig.OpenIdAuthorization, "/authorize", "/userinfo", -1)
-		log.Printf("Userinfo endpoint: %s", userinfoEndpoint)
-		client := &http.Client{}
-		req, err := http.NewRequest(
-			"GET",
-			userinfoEndpoint,
-			nil,
-		)
-
-		//req.Header.Add("accept", "application/json")
-		//req.Header.Add("cache-control", "no-cache")
-		req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", openid.AccessToken))
-		res, err := client.Do(req)
-		if err != nil {
-			log.Printf("[WARNING] OpenID client DO (2): %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false, "reason": "Failed userinfo request"}`))
-			return
-		}
-
-		body, err = ioutil.ReadAll(res.Body)
-		if err != nil {
-			log.Printf("[WARNING] OpenID client Body (2): %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false, "reason": "Failed userinfo body parsing"}`))
-			return
-		}
-
-		err = json.Unmarshal(body, &openidUser)
-		if err != nil {
-			log.Printf("[WARNING] Error in Openid marshal (2): %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-	}
-
-	//log.Printf("Got user body: %s", string(body))
-
-	/*
-
-		BELOW HERE ITS ALL COPY PASTE OF USER INFO THINGS!
-
-	*/
-
-	if len(openidUser.Sub) == 0 {
-		log.Printf("[WARNING] No user found in openid login (2)")
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	if project.Environment == "cloud" {
-		log.Printf("[WARNING] Openid SSO is not implemented for cloud yet. User %s", openidUser.Sub)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Cloud Openid is not available yet"}`)))
-		return
-	}
-
-	userName := openidUser.Sub
-	redirectUrl := "/workflows"
-
-	users, err := FindGeneratedUser(ctx, strings.ToLower(strings.TrimSpace(userName)))
-	if err == nil && len(users) > 0 {
-		for _, user := range users {
-			log.Printf("%s - %s", user.GeneratedUsername, userName)
-			if user.GeneratedUsername == userName {
-				log.Printf("[AUDIT] Found user %s (%s) which matches SSO info for %s. Redirecting to login!", user.Username, user.Id, userName)
-
-				//log.Printf("SESSION: %s", user.Session)
-
-				expiration := time.Now().Add(3600 * time.Second)
-				//if len(user.Session) == 0 {
-				log.Printf("[INFO] User does NOT have session - creating")
-				sessionToken := uuid.NewV4().String()
-
-				newCookie := http.Cookie{
-					Name:    "session_token",
-					Value:   sessionToken,
-					Expires: expiration,
-				}
-
-				if project.Environment == "cloud" {
-					newCookie.Domain = ".shuffler.io"
-					newCookie.Secure = true
-					newCookie.HttpOnly = true
-				}
-
-				http.SetCookie(resp, &newCookie)
-
-				newCookie.Name = "__session"
-				http.SetCookie(resp, &newCookie)
-
-				err = SetSession(ctx, user, sessionToken)
-				if err != nil {
-					log.Printf("[WARNING] Error creating session for user: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-					return
-				}
-
-				user.Session = sessionToken
-				err = SetUser(ctx, &user, false)
-				if err != nil {
-					log.Printf("[WARNING] Failed updating user when setting session: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(`{"success": false, "reason": "Failed user update during session storage (2)"}`))
-					return
-				}
-
-				//redirectUrl = fmt.Sprintf("%s?source=SSO&id=%s", redirectUrl, session)
-				http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-				return
-			}
-		}
-	}
-
-	// Normal user. Checking because of backwards compatibility. Shouldn't break anything as we have unique names
-	users, err = FindUser(ctx, strings.ToLower(strings.TrimSpace(userName)))
-	if err == nil && len(users) > 0 {
-		for _, user := range users {
-			if user.Username == userName {
-				log.Printf("[AUDIT] Found user %s (%s) which matches SSO info for %s. Redirecting to login %s!", user.Username, user.Id, userName, redirectUrl)
-
-				//log.Printf("SESSION: %s", user.Session)
-
-				expiration := time.Now().Add(3600 * time.Second)
-				//if len(user.Session) == 0 {
-				log.Printf("[INFO] User does NOT have session - creating")
-				sessionToken := uuid.NewV4().String()
-				newCookie := &http.Cookie{
-					Name:    "session_token",
-					Value:   sessionToken,
-					Expires: expiration,
-				}
-
-				if project.Environment == "cloud" {
-					newCookie.Domain = ".shuffler.io"
-					newCookie.Secure = true
-					newCookie.HttpOnly = true
-				}
-
-				http.SetCookie(resp, newCookie)
-
-				newCookie.Name = "__session"
-				http.SetCookie(resp, newCookie)
-
-				err = SetSession(ctx, user, sessionToken)
-				if err != nil {
-					log.Printf("[WARNING] Error creating session for user: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-					return
-				}
-
-				user.Session = sessionToken
-				err = SetUser(ctx, &user, false)
-				if err != nil {
-					log.Printf("[WARNING] Failed updating user when setting session: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(`{"success": false, "reason": "Failed user update during session storage (2)"}`))
-					return
-				}
-
-				//redirectUrl = fmt.Sprintf("%s?source=SSO&id=%s", redirectUrl, session)
-				http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-				return
-			}
-		}
-	}
-
-	/*
-		orgs, err := GetAllOrgs(ctx)
-		if err != nil {
-			log.Printf("[WARNING] Failed finding orgs during SSO setup: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed getting valid organizations"}`)))
-			return
-		}
-
-		foundOrg := Org{}
-		for _, org := range orgs {
-			if len(org.ManagerOrgs) == 0 {
-				foundOrg = org
-				break
-			}
-		}
-	*/
-
-	if len(org.Id) == 0 {
-		log.Printf("[WARNING] Failed finding a valid org (default) without suborgs during SSO setup")
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed finding valid SSO auto org"}`)))
-		return
-	}
-
-	log.Printf("[AUDIT] Adding user %s to org %s (%s) through single sign-on", userName, org.Name, org.Id)
-	newUser := new(User)
-	// Random password to ensure its not empty
-	newUser.Password = uuid.NewV4().String()
-	newUser.Username = userName
-	newUser.GeneratedUsername = userName
-	newUser.Verified = true
-	newUser.Active = true
-	newUser.CreationTime = time.Now().Unix()
-	newUser.Orgs = []string{org.Id}
-	newUser.LoginType = "OpenID"
-	newUser.Role = "user"
-	newUser.Session = uuid.NewV4().String()
-
-	verifyToken := uuid.NewV4()
-	ID := uuid.NewV4()
-	newUser.Id = ID.String()
-	newUser.VerificationToken = verifyToken.String()
-
-	expiration := time.Now().Add(3600 * time.Second)
-	//if len(user.Session) == 0 {
-	log.Printf("[INFO] User does NOT have session - creating")
-	sessionToken := uuid.NewV4().String()
-
-	newCookie := &http.Cookie{
-		Name:    "session_token",
-		Value:   sessionToken,
-		Expires: expiration,
-	}
-
-	if project.Environment == "cloud" {
-		newCookie.Domain = ".shuffler.io"
-		newCookie.Secure = true
-		newCookie.HttpOnly = true
-	}
-
-	http.SetCookie(resp, newCookie)
-
-	newCookie.Name = "__session"
-	http.SetCookie(resp, newCookie)
-
-	err = SetSession(ctx, *newUser, sessionToken)
-	if err != nil {
-		log.Printf("[WARNING] Error creating session for user: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-		return
-	}
-
-	newUser.Session = sessionToken
-	err = SetUser(ctx, newUser, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting new user in DB: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed updating the user"}`)))
-		return
-	}
-
-	http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-	return
-}
-
-// Example implementation of SSO, including a redirect for the user etc
-// Should make this stuff only possible after login
-func HandleSSO(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	//log.Printf("SSO LOGIN: %#v", request)
-	// Deserialize
-	// Serialize
-
-	// SAML
-	//entryPoint := "https://dev-23367303.okta.com/app/dev-23367303_shuffletest_1/exk1vg1j7bYUYEG0k5d7/sso/saml"
-	redirectUrl := "http://localhost:3001/workflows"
-	backendUrl := os.Getenv("SSO_REDIRECT_URL")
-	if len(backendUrl) == 0 && len(os.Getenv("BASE_URL")) > 0 {
-		backendUrl = os.Getenv("BASE_URL")
-	}
-
-	if len(backendUrl) > 0 {
-		redirectUrl = fmt.Sprintf("%s/workflows", backendUrl)
-	}
-
-	if project.Environment == "cloud" {
-		redirectUrl = "https://shuffler.io/workflows"
-
-		if len(os.Getenv("SHUFFLE_GCEPROJECT")) > 0 && len(os.Getenv("SHUFFLE_GCEPROJECT_LOCATION")) > 0 {
-			backendUrl = fmt.Sprintf("https://%s.%s.r.appspot.com/workflows", os.Getenv("SHUFFLE_GCEPROJECT"), os.Getenv("SHUFFLE_GCEPROJECT_LOCATION"))
-		}
-
-		if len(os.Getenv("SHUFFLE_CLOUDRUN_URL")) > 0 {
-			backendUrl = os.Getenv("SHUFFLE_CLOUDRUN_URL")
-		}
-	}
-
-	log.Printf("[DEBUG] Using %s as redirectUrl in SSO", backendUrl)
-
-	//backendUrl := os.Getenv("BASE_URL")
-	//if project.Environment == "cloud" {
-	//	backendUrl = "https://shuffler.io"
-	//}
-	//if len(backendUrl) == 0 {
-	//	backendUrl = "http://127.0.0.1:5001"
-	//}
-
-	//log.Printf("URL: %#v", request.URL)
-	//log.Printf("REDIRECT: %s", redirectUrl)
-	//_ = entryPoint
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("[WARNING] Error with body read of SSO: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Parsing out without using Field
-	// This is a mess, but all made to handle base64 and equal signs
-	parsedSAML := ""
-	for _, item := range strings.Split(string(body), "&") {
-		//log.Printf("Got body with info: %s", item)
-		if strings.Contains(item, "SAMLRequest") || strings.Contains(item, "SAMLResponse") {
-			equalsplit := strings.Split(item, "=")
-			addedEquals := len(equalsplit)
-			if len(equalsplit) >= 2 {
-				//bareEquals := strings.Join(equalsplit[1:len(equalsplit)-1], "=")
-				bareEquals := equalsplit[1]
-				//log.Printf("Equal: %s", bareEquals)
-				_ = addedEquals
-				//if len(strings.Split(bareEquals, "=")) < addedEquals {
-				//	bareEquals += "="
-				//}
-
-				decodedValue, err := url.QueryUnescape(bareEquals)
-				if err != nil {
-					log.Printf("[WARNING] Failed url query escape: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed decoding saml value"}`)))
-					return
-				}
-
-				if strings.Contains(decodedValue, " ") {
-					decodedValue = strings.Replace(decodedValue, " ", "+", -1)
-				}
-
-				parsedSAML = decodedValue
-				break
-			}
-		}
-	}
-
-	if len(parsedSAML) == 0 {
-		log.Printf("[WARNING] No SAML to be parsed from request.")
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "No data to parse. Is the request missing the SAMLResponse query?"}`)))
-		return
-	}
-
-	bytesXML, err := base64.StdEncoding.DecodeString(parsedSAML)
-	if err != nil {
-		log.Printf("[WARNING] Failed base64 decode of SAML: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed base64 decoding in SAML"}`)))
-		return
-	}
-
-	//log.Printf("Parsed: %s", bytesXML)
-
-	// Sample request in keycloak lab env
-	// PS: Should it ever come this way..?
-	//<samlp:AuthnRequest xmlns:samlp="urn:oasis:names:tc:SAML:2.0:protocol" xmlns="urn:oasis:names:tc:SAML:2.0:assertion" xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion" AssertionConsumerServiceURL="http://192.168.55.2:8080/auth/realms/ShuffleSSOSaml/broker/shaffuru/endpoint" Destination="http://192.168.55.2:3001/api/v1/login_sso" ForceAuthn="false" ID="" IssueInstant="2022-01-31T20:24:37.238Z" ProtocolBinding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Version="2.0"><saml:Issuer>http://192.168.55.2:8080/auth/realms/ShuffleSSOSaml</saml:Issuer><samlp:NameIDPolicy AllowCreate="false" Format="urn:oasis:names:tc:SAML:1.1:nameid-format:emailAddress"/></samlp:AuthnRequest>
-
-	var samlResp SAMLResponse
-	err = xml.Unmarshal(bytesXML, &samlResp)
-	if err != nil {
-		if strings.Contains(fmt.Sprintf("%s", err), "AuthnRequest") {
-			var newSamlResp SamlRequest
-			err = xml.Unmarshal(bytesXML, &newSamlResp)
-			if err != nil {
-				log.Printf("[WARNING] Failed XML unmarshal (2): %s", err)
-				resp.WriteHeader(401)
-				resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed XML unpacking in SAML (2)"}`)))
-				return
-			}
-
-			// Being here means we need to redirect
-			log.Printf("[DEBUG] Handling authnrequest redirect back to %s? That's not how any of this works.", newSamlResp.AssertionConsumerServiceURL)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed XML unpacking in SAML (2)"}`)))
-			return
-
-			// User tries to access a protected resource on the SP. SP checks if the user has a local (and authenticated session). If not it generates a SAML <AuthRequest> which includes a random id. The SP then redirects the user to the IDP with this AuthnRequest.
-
-			return
-		} else {
-			log.Printf("[WARNING] Failed XML unmarshal: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed XML unpacking in SAML (1)"}`)))
-			return
-		}
-	}
-
-	baseCertificate := samlResp.Signature.KeyInfo.X509Data.X509Certificate
-	if len(baseCertificate) == 0 {
-		//log.Printf("%#v", samlResp.Signature.KeyInfo.X509Data)
-		baseCertificate = samlResp.Assertion.Signature.KeyInfo.X509Data.X509Certificate
-	}
-
-	//log.Printf("\n\n%d - CERT: %s\n\n", len(baseCertificate), baseCertificate)
-	parsedX509Key := fixCertificate(baseCertificate)
-
-	ctx := GetContext(request)
-	matchingOrgs, err := GetOrgByField(ctx, "sso_config.sso_certificate", parsedX509Key)
-	if err != nil {
-		log.Printf("[DEBUG] BYTES FROM REQUEST (DEBUG): %s", string(bytesXML))
-
-		log.Printf("[WARNING] Bad certificate (%d): Failed to find a org with certificate matching the SSO", len(parsedX509Key))
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed finding an org with the right certificate"}`)))
-		return
-	}
-
-	// Validating the orgs
-	if len(matchingOrgs) >= 1 {
-		newOrgs := []Org{}
-		for _, org := range matchingOrgs {
-			if org.SSOConfig.SSOCertificate == parsedX509Key {
-				newOrgs = append(newOrgs, org)
-			} else {
-				log.Printf("[WARNING] Skipping org append because bad cert: %d vs %d", len(org.SSOConfig.SSOCertificate), len(parsedX509Key))
-
-				//log.Printf(parsedX509Key)
-				//log.Printf(org.SSOConfig.SSOCertificate)
-			}
-		}
-
-		matchingOrgs = newOrgs
-	}
-
-	if len(matchingOrgs) != 1 {
-		log.Printf("[DEBUG] BYTES FROM REQUEST (2 - DEBUG): %s", string(bytesXML))
-		log.Printf("[WARNING] Bad certificate (%d). Original orgs: %d: X509 doesnt match certificate for any organization", len(parsedX509Key), len(matchingOrgs))
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Certificate for SSO doesn't match any organization"}`)))
-		return
-	}
-
-	foundOrg := matchingOrgs[0]
-	userName := samlResp.Assertion.Subject.NameID.Text
-	if len(userName) == 0 {
-		log.Printf("[WARNING] Failed finding user - No name: %#v", samlResp.Assertion.Subject)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed finding a user to authenticate"}`)))
-		return
-	}
-
-	// Start actually fixing the user
-	// 1. Check if the user exists - if it does - give it a valid cookie
-	// 2. If it doesn't, find the correct org to connect them with, then register them
-
-	/*
-		if project.Environment == "cloud" {
-			log.Printf("[WARNING] SAML SSO is not implemented for cloud yet")
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Cloud SSO is not available for you"}`)))
-			return
-		}
-	*/
-
-	users, err := FindGeneratedUser(ctx, strings.ToLower(strings.TrimSpace(userName)))
-	if err == nil && len(users) > 0 {
-		for _, user := range users {
-			log.Printf("%s - %s", user.GeneratedUsername, userName)
-			if user.GeneratedUsername == userName {
-				log.Printf("[AUDIT] Found user %s (%s) which matches SSO info for %s. Redirecting to login!", user.Username, user.Id, userName)
-
-				if project.Environment == "cloud" {
-					user.ActiveOrg.Id = matchingOrgs[0].Id
-				}
-
-				//log.Printf("SESSION: %s", user.Session)
-
-				expiration := time.Now().Add(3600 * time.Second)
-				//if len(user.Session) == 0 {
-				log.Printf("[INFO] User does NOT have session - creating")
-				sessionToken := uuid.NewV4().String()
-				newCookie := &http.Cookie{
-					Name:    "session_token",
-					Value:   sessionToken,
-					Expires: expiration,
-				}
-
-				if project.Environment == "cloud" {
-					newCookie.Domain = ".shuffler.io"
-					newCookie.Secure = true
-					newCookie.HttpOnly = true
-				}
-
-				http.SetCookie(resp, newCookie)
-
-				newCookie.Name = "__session"
-				http.SetCookie(resp, newCookie)
-
-				err = SetSession(ctx, user, sessionToken)
-				if err != nil {
-					log.Printf("[WARNING] Error creating session for user: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-					return
-				}
-
-				user.Session = sessionToken
-				err = SetUser(ctx, &user, false)
-				if err != nil {
-					log.Printf("[WARNING] Failed updating user when setting session: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(`{"success": false, "reason": "Failed user update during session storage (2)"}`))
-					return
-				}
-
-				//redirectUrl = fmt.Sprintf("%s?source=SSO&id=%s", redirectUrl, session)
-				http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-				return
-			}
-		}
-	}
-
-	// Normal user. Checking because of backwards compatibility. Shouldn't break anything as we have unique names
-	users, err = FindUser(ctx, strings.ToLower(strings.TrimSpace(userName)))
-	if err == nil && len(users) > 0 {
-		for _, user := range users {
-			if user.Username == userName {
-				log.Printf("[AUDIT] Found user %s (%s) which matches SSO info for %s. Redirecting to login %s!", user.Username, user.Id, userName, redirectUrl)
-
-				//log.Printf("SESSION: %s", user.Session)
-				if project.Environment == "cloud" {
-					user.ActiveOrg.Id = matchingOrgs[0].Id
-				}
-
-				expiration := time.Now().Add(3600 * time.Second)
-				//if len(user.Session) == 0 {
-				log.Printf("[INFO] User does NOT have session - creating")
-				sessionToken := uuid.NewV4().String()
-				newCookie := &http.Cookie{
-					Name:    "session_token",
-					Value:   sessionToken,
-					Expires: expiration,
-				}
-
-				if project.Environment == "cloud" {
-					newCookie.Domain = ".shuffler.io"
-					newCookie.Secure = true
-					newCookie.HttpOnly = true
-				}
-
-				http.SetCookie(resp, newCookie)
-
-				newCookie.Name = "__session"
-				http.SetCookie(resp, newCookie)
-
-				err = SetSession(ctx, user, sessionToken)
-				if err != nil {
-					log.Printf("[WARNING] Error creating session for user: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-					return
-				}
-
-				user.Session = sessionToken
-				err = SetUser(ctx, &user, false)
-				if err != nil {
-					log.Printf("[WARNING] Failed updating user when setting session: %s", err)
-					resp.WriteHeader(401)
-					resp.Write([]byte(`{"success": false, "reason": "Failed user update during session storage (2)"}`))
-					return
-				}
-
-				//redirectUrl = fmt.Sprintf("%s?source=SSO&id=%s", redirectUrl, session)
-				http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-				return
-			}
-		}
-	}
-
-	/*
-		orgs, err := GetAllOrgs(ctx)
-		if err != nil {
-			log.Printf("[WARNING] Failed finding orgs during SSO setup: %s", err)
-			resp.WriteHeader(401)
-			resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed getting valid organizations"}`)))
-			return
-		}
-
-		foundOrg := Org{}
-		for _, org := range orgs {
-			if len(org.ManagerOrgs) == 0 {
-				foundOrg = org
-				break
-			}
-		}
-	*/
-
-	if len(foundOrg.Id) == 0 {
-		log.Printf("[WARNING] Failed finding a valid org (default) without suborgs during SSO setup")
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed finding valid SSO auto org"}`)))
-		return
-	}
-
-	log.Printf("[AUDIT] Adding user %s to org %s (%s) through single sign-on", userName, foundOrg.Name, foundOrg.Id)
-	newUser := new(User)
-	// Random password to ensure its not empty
-	newUser.Password = uuid.NewV4().String()
-	newUser.Username = userName
-	newUser.GeneratedUsername = userName
-	newUser.Verified = true
-	newUser.Active = true
-	newUser.CreationTime = time.Now().Unix()
-	newUser.Orgs = []string{foundOrg.Id}
-	newUser.LoginType = "SSO"
-	newUser.Role = "user"
-	newUser.Session = uuid.NewV4().String()
-
-	newUser.ActiveOrg.Id = matchingOrgs[0].Id
-
-	verifyToken := uuid.NewV4()
-	ID := uuid.NewV4()
-	newUser.Id = ID.String()
-	newUser.VerificationToken = verifyToken.String()
-
-	expiration := time.Now().Add(3600 * time.Second)
-	//if len(user.Session) == 0 {
-	log.Printf("[INFO] User does NOT have session - creating")
-	sessionToken := uuid.NewV4().String()
-	newCookie := &http.Cookie{
-		Name:    "session_token",
-		Value:   sessionToken,
-		Expires: expiration,
-	}
-
-	if project.Environment == "cloud" {
-		newCookie.Domain = ".shuffler.io"
-		newCookie.Secure = true
-		newCookie.HttpOnly = true
-	}
-
-	http.SetCookie(resp, newCookie)
-
-	newCookie.Name = "__session"
-	http.SetCookie(resp, newCookie)
-
-	err = SetSession(ctx, *newUser, sessionToken)
-	if err != nil {
-		log.Printf("[WARNING] Error creating session for user: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed setting session"}`)))
-		return
-	}
-
-	newUser.Session = sessionToken
-	err = SetUser(ctx, newUser, true)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting new user in DB: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed updating the user"}`)))
-		return
-	}
-
-	http.Redirect(resp, request, redirectUrl, http.StatusSeeOther)
-	return
-}
-
-// Downloads documentation from Github to be placed in an app/workflow as markdown
-// Caching no matter what, with no retries
-func DownloadFromUrl(ctx context.Context, url string) ([]byte, error) {
-	cacheKey := fmt.Sprintf("docs_%s", url)
-	cache, err := GetCache(ctx, cacheKey)
-	if err == nil {
-		cacheData := []byte(cache.([]uint8))
-		return cacheData, nil
-	}
-
-	httpClient := &http.Client{}
-	req, err := http.NewRequest(
-		"GET",
-		url,
-		nil,
-	)
-
-	if err != nil {
-		SetCache(ctx, cacheKey, []byte{})
-		return []byte{}, err
-	}
-
-	newresp, err := httpClient.Do(req)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	//log.Printf("URL %#v, RESP: %d", url, newresp.StatusCode)
-	if newresp.StatusCode != 200 {
-		SetCache(ctx, cacheKey, []byte{})
-
-		return []byte{}, errors.New(fmt.Sprintf("No body to handle for %#v. Status: %d", url, newresp.StatusCode))
-	}
-
-	body, err := ioutil.ReadAll(newresp.Body)
-	if err != nil {
-		SetCache(ctx, cacheKey, []byte{})
-		return []byte{}, err
-	}
-
-	//log.Printf("Documentation: %#v", string(body))
-	if len(body) > 0 {
-		err = SetCache(ctx, cacheKey, body)
-		if err != nil {
-			log.Printf("[WARNING] Failed setting cache for workflow/app doc %s: %s", url, err)
-		}
-		return body, nil
-	}
-
-	SetCache(ctx, cacheKey, []byte{})
-	return []byte{}, errors.New(fmt.Sprintf("No body to handle for %#v", url))
-}
-
 func HealthCheckHandler(resp http.ResponseWriter, request *http.Request) {
 	ret, err := project.Es.Info()
 	if err != nil {
@@ -5146,245 +1847,6 @@ func HealthCheckHandler(resp http.ResponseWriter, request *http.Request) {
 	}
 
 	fmt.Fprint(resp, "OK")
-}
-
-// Extra validation sample to be used for workflow executions based on parent workflow instead of users' auth
-
-// Check if the execution data has correct info in it! Happens based on subflows.
-// 1. Parent workflow contains this workflow ID in the source trigger?
-// 2. Parent workflow's owner is same org?
-// 3. Parent execution auth is correct
-func RunExecuteAccessValidation(request *http.Request, workflow *Workflow) (bool, string) {
-	log.Printf("[DEBUG] Inside execute validation!")
-
-	if request.Method == "POST" {
-		ctx := GetContext(request)
-		workflowExecution := &WorkflowExecution{}
-		sourceExecution, sourceExecutionOk := request.URL.Query()["source_execution"]
-		if sourceExecutionOk {
-			//log.Printf("[DEBUG] Got source exec %s", sourceExecution)
-			newExec, err := GetWorkflowExecution(ctx, sourceExecution[0])
-			if err != nil {
-				log.Printf("[INFO] Failed getting source_execution in test validation based on %#v", sourceExecution[0])
-				return false, ""
-			} else {
-				workflowExecution = newExec
-			}
-
-		} else {
-			return false, ""
-		}
-
-		if workflowExecution.ExecutionId == "" {
-			log.Printf("[WARNING] No execution ID found. Bad auth.")
-			return false, ""
-		}
-
-		sourceAuth, sourceAuthOk := request.URL.Query()["source_auth"]
-		if sourceAuthOk {
-			//log.Printf("[DEBUG] Got auth %s", sourceAuth)
-
-			if sourceAuth[0] != workflowExecution.Authorization {
-				log.Printf("[WARNING] Bad authorization for workflowexecution defined.")
-				return false, ""
-			}
-		} else {
-			return false, ""
-		}
-
-		// When reaching here, authentication is done, but not authorization.
-		// Need to verify the workflow, and whether it SHOULD have access to execute it.
-		sourceWorkflow, sourceWorkflowOk := request.URL.Query()["source_workflow"]
-		if sourceWorkflowOk {
-			//log.Printf("[DEBUG] Got source workflow %s", sourceWorkflow)
-			_ = sourceWorkflow
-
-			// Source workflow = parent
-			// This workflow = child
-
-			//if sourceWorkflow[0] != workflow.ID {
-			//	log.Printf("[DEBUG] Bad workflow in execution.")
-			//	return false, ""
-			//}
-
-		} else {
-			//log.Printf("Did NOT get source workflow")
-			return false, ""
-		}
-
-		if workflow.OrgId != workflowExecution.Workflow.OrgId || workflow.ExecutingOrg.Id != workflowExecution.Workflow.ExecutingOrg.Id || workflow.OrgId == "" {
-			//e9274e37e53631a2321747b1be088f4d2631a6300a309eec9b4515c8528c35f4
-			return false, ""
-		}
-
-		// 1. Parent workflow contains this workflow ID in the source trigger?
-		// 2. Parent workflow's owner is same org?
-		// 3. Parent execution auth is correct
-
-		sourceNode, sourceNodeOk := request.URL.Query()["source_node"]
-		if sourceNodeOk {
-			//log.Printf("[DEBUG] Got source node %s", sourceNode)
-			//workflowExecution.ExecutionSourceNode = sourceNode[0]
-		} else {
-			return false, ""
-		}
-
-		// SHOULD be executed by a trigger in the parent.
-		for _, trigger := range workflowExecution.Workflow.Triggers {
-			if sourceNode[0] == trigger.ID {
-				return true, workflowExecution.ExecutionOrg
-			}
-		}
-	}
-
-	return false, ""
-}
-
-// Significantly slowed down everything. Just returning for now.
-func findReferenceAppDocs(ctx context.Context, allApps []WorkflowApp) []WorkflowApp {
-	newApps := []WorkflowApp{}
-
-	// Skipping this for now as it makes things slow
-	return allApps
-
-	for _, app := range allApps {
-		if len(app.ReferenceInfo.DocumentationUrl) > 0 && strings.HasPrefix(app.ReferenceInfo.DocumentationUrl, "https://raw.githubusercontent.com/Shuffle") && strings.Contains(app.ReferenceInfo.DocumentationUrl, ".md") {
-			// Should find documentation from the github (only if github?) and add it to app.Documentation before caching
-			//log.Printf("DOCS: %#v", app.ReferenceInfo.DocumentationUrl)
-			documentationData, err := DownloadFromUrl(ctx, app.ReferenceInfo.DocumentationUrl)
-			if err != nil {
-				log.Printf("[ERROR] Failed getting data: %#v", err)
-			} else {
-				app.Documentation = string(documentationData)
-			}
-		}
-
-		//if app.Documentation == "" && strings.ToLower(app.Name) == "discord" {
-		if app.Documentation == "" {
-			referenceUrl := ""
-
-			if app.Generated {
-				//log.Printf("[DEBUG] Should look in the OpenAPI folder")
-				baseUrl := "https://raw.githubusercontent.com/Shuffle/openapi-apps/master/docs"
-
-				newName := strings.ToLower(strings.Replace(strings.Replace(app.Name, " ", "_", -1), "-", "_", -1))
-				referenceUrl = fmt.Sprintf("%s/%s.md", baseUrl, newName)
-			} else {
-				//log.Printf("[DEBUG] Should look in the Python-apps folder")
-			}
-
-			if len(referenceUrl) > 0 {
-				//log.Printf("REF: %#v", referenceUrl)
-
-				documentationData, err := DownloadFromUrl(ctx, referenceUrl)
-				if err != nil {
-					log.Printf("[ERROR] Failed getting documentation data for app %#v: %#v", app.Name, err)
-				} else {
-					//log.Printf("[INFO] Added documentation from github for %#v", app.Name)
-					app.Documentation = string(documentationData)
-				}
-			}
-		}
-
-		newApps = append(newApps, app)
-	}
-
-	return newApps
-}
-
-func EchoOpenapiData(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	// Just here to verify that the user is logged in
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[DEBUG] Api authentication failed in download Yaml echo: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed authentication"}`))
-		return
-	}
-
-	if user.Role == "org-reader" {
-		log.Printf("[WARNING] Org-reader doesn't have access to echo OpenAPI data: %s (%s)", user.Username, user.Id)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Read only user"}`))
-		return
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("Bodyreader err: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed reading body"}`))
-		return
-	}
-
-	newbody := string(body)
-	newbody = strings.TrimSpace(newbody)
-	if strings.HasPrefix(newbody, "\"") {
-		newbody = newbody[1:len(newbody)]
-	}
-
-	if strings.HasSuffix(newbody, "\"") {
-		newbody = newbody[0 : len(newbody)-1]
-	}
-
-	// Rewrite to download proper one from Github even without raw path
-	if strings.Contains(newbody, "https://github.com/") {
-		// https://github.com/AdguardTeam/AdGuardHome/blob/master/openapi/openapi.yaml
-		// https://raw.githubusercontent.com/AdguardTeam/AdGuardHome/master/openapi/openapi.yaml
-		// https://raw.githubusercontent.com/AdguardTeam/AdGuardHome/master/openapi/openapi.yaml
-
-		urlsplit := strings.Split(newbody, "/")
-		if len(urlsplit) > 6 {
-			log.Printf("[DEBUG] Rewriting github URL %s.", newbody)
-			ghuser := urlsplit[3]
-			repo := urlsplit[4]
-			branch := urlsplit[6]
-			path := strings.Join(urlsplit[7:len(urlsplit)], "/")
-			newbody = fmt.Sprintf("https://raw.githubusercontent.com/%s/%s/%s/%s", ghuser, repo, branch, path)
-		}
-	}
-
-	log.Printf("[DEBUG] Downloading content from %s", newbody)
-
-	req, err := http.NewRequest("GET", newbody, nil)
-	if err != nil {
-		log.Printf("[ERROR] Requestbuilder err: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(`{"success": false, "reason": "Failed building request"}`))
-		return
-	}
-
-	httpClient := &http.Client{}
-	newresp, err := httpClient.Do(req)
-	if err != nil {
-		log.Printf("[ERROR] Grabbing error: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Failed making remote request to get the data"}`)))
-		return
-	}
-	defer newresp.Body.Close()
-
-	urlbody, err := ioutil.ReadAll(newresp.Body)
-	if err != nil {
-		log.Printf("[ERROR] URLbody error: %s", err)
-		resp.WriteHeader(500)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "Can't get data from selected uri"}`)))
-		return
-	}
-
-	if newresp.StatusCode >= 400 {
-		resp.WriteHeader(201)
-		resp.Write([]byte(fmt.Sprintf(`{"success": false, "reason": "%s"}`, urlbody)))
-		return
-	}
-
-	resp.WriteHeader(200)
-	resp.Write(urlbody)
 }
 
 func GetFrameworkConfiguration(resp http.ResponseWriter, request *http.Request) {
@@ -5604,196 +2066,6 @@ func (fw *flushWriter) Write(p []byte) (n int, err error) {
 		fw.f.Flush()
 	}
 	return
-}
-
-func HandleStreamWorkflow(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	//// Removed check here as it may be a public workflow
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[AUDIT] Api authentication failed in getting specific workflow (stream): %s. Continuing because it may be public.", err)
-	}
-
-	location := strings.Split(request.URL.String(), "/")
-
-	var fileId string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if strings.Contains(fileId, "?") {
-		fileId = strings.Split(fileId, "?")[0]
-	}
-
-	if len(fileId) != 36 {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Workflow ID when getting workflow is not valid"}`))
-		return
-	}
-
-	//ctx := GetContext(request)
-	ctx := context.Background()
-	workflow, err := GetWorkflow(ctx, fileId)
-	if err != nil {
-		log.Printf("[WARNING] Workflow %s doesn't exist.", fileId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed finding workflow."}`))
-		return
-	}
-
-	if user.Id != workflow.Owner || len(user.Id) == 0 {
-		if workflow.OrgId == user.ActiveOrg.Id && (user.Role == "admin" || user.Role == "org-reader") {
-			log.Printf("[AUDIT] User %s is accessing workflow %s as admin (stream edit workflow)", user.Username, workflow.ID)
-		} else if workflow.Public {
-			log.Printf("[AUDIT] Letting user %#v access workflow %s for streaming because it's public", user.Username, workflow.ID)
-		} else if project.Environment == "cloud" && user.Verified == true && user.Active == true && user.SupportAccess == true && strings.HasSuffix(user.Username, "@shuffler.io") {
-			log.Printf("[AUDIT] Letting verified support admin %s access workflow %s", user.Username, workflow.ID)
-		} else {
-			log.Printf("[AUDIT] Wrong user (%s) for workflow %s (get workflow)", user.Username, workflow.ID)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-	}
-
-	resp.Header().Set("Connection", "Keep-Alive")
-	resp.Header().Set("X-Content-Type-Options", "nosniff")
-
-	conn, ok := resp.(http.Flusher)
-	if !ok {
-		log.Printf("[ERROR] Flusher error: %s", ok)
-		http.Error(resp, "Streaming supported on AppEngine", http.StatusInternalServerError)
-		return
-	}
-
-	resp.Header().Set("Content-Type", "text/event-stream")
-	resp.WriteHeader(http.StatusOK)
-
-	sessionKey := fmt.Sprintf("%s_stream", workflow.ID)
-	previousCache := []byte{}
-	for {
-		cache, err := GetCache(ctx, sessionKey)
-		if err == nil {
-			cacheData := []byte(cache.([]uint8))
-			if string(previousCache) == string(cacheData) {
-				//log.Printf("[DEBUG] Still same cache for %#v", user.Id)
-			} else {
-				// A way to only check for data from other people
-				if (len(user.Id) > 0 && !strings.Contains(string(cacheData), user.Id)) || len(user.Id) == 0 {
-					//fw.Write(cacheData)
-					//w.Write(cacheData)
-
-					_, err := fmt.Fprintf(resp, string(cacheData))
-					if err != nil {
-						log.Printf("[ERROR] Failed in writing stream to user: %s", err)
-					} else {
-						previousCache = cacheData
-						conn.Flush()
-					}
-
-				} else {
-					//log.Printf("[ERROR] NEW cache for %#v (2) - NOT sending: %s.", user.Id, cacheData)
-
-					previousCache = cacheData
-				}
-
-			}
-		} else {
-			//log.Printf("[DEBUG] Failed getting cache for %#v: %s", user.Id, err)
-		}
-
-		time.Sleep(500 * time.Millisecond)
-	}
-}
-
-func HandleStreamWorkflowUpdate(resp http.ResponseWriter, request *http.Request) {
-	cors := HandleCors(resp, request)
-	if cors {
-		return
-	}
-
-	//// Removed check here as it may be a public workflow
-	user, err := HandleApiAuthentication(resp, request)
-	if err != nil {
-		log.Printf("[AUDIT] Api authentication failed in getting specific workflow (stream update): %s. Continuing because it may be public.", err)
-	}
-
-	location := strings.Split(request.URL.String(), "/")
-
-	var fileId string
-	if location[1] == "api" {
-		if len(location) <= 4 {
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-
-		fileId = location[4]
-	}
-
-	if strings.Contains(fileId, "?") {
-		fileId = strings.Split(fileId, "?")[0]
-	}
-
-	if len(fileId) != 36 {
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Workflow ID when getting workflow is not valid"}`))
-		return
-	}
-
-	ctx := GetContext(request)
-	workflow, err := GetWorkflow(ctx, fileId)
-	if err != nil {
-		log.Printf("[WARNING] Workflow %s doesn't exist.", fileId)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false, "reason": "Failed finding workflow."}`))
-		return
-	}
-
-	if user.Id != workflow.Owner || len(user.Id) == 0 {
-		if workflow.OrgId == user.ActiveOrg.Id && (user.Role == "admin" || user.Role == "org-reader") {
-			log.Printf("[AUDIT] User %s is accessing workflow %s as admin (get workflow)", user.Username, workflow.ID)
-
-		} else if project.Environment == "cloud" && user.Verified == true && user.SupportAccess == true && user.Role == "admin" {
-			log.Printf("[AUDIT] Letting verified support admin %s access workflow %s", user.Username, workflow.ID)
-
-		} else {
-			log.Printf("[AUDIT] Wrong user (%s) for workflow %s (get workflow stream)", user.Username, workflow.ID)
-			resp.WriteHeader(401)
-			resp.Write([]byte(`{"success": false}`))
-			return
-		}
-	}
-
-	body, err := ioutil.ReadAll(request.Body)
-	if err != nil {
-		log.Printf("[WARNING] Error with body read in workflow stream: %s", err)
-		resp.WriteHeader(401)
-		resp.Write([]byte(`{"success": false}`))
-		return
-	}
-
-	// Literally just dumping them in, as they're supposed to be overwritten continuously
-	// PS: This is NOT an ideal process, and broadcasting should be handled differently
-	//log.Printf("Body to update: %s", string(body))
-	sessionKey := fmt.Sprintf("%s_stream", workflow.ID)
-	err = SetCache(ctx, sessionKey, body)
-	if err != nil {
-		log.Printf("[WARNING] Failed setting cache for apikey: %s", err)
-	}
-
-	resp.WriteHeader(200)
-	resp.Write([]byte("OK"))
 }
 
 func LoadUsecases(resp http.ResponseWriter, request *http.Request) {
